% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dpp.R
\name{addAge}
\alias{addAge}
\title{Add comorbidity}
\usage{
addAge(x, target, at, ...)
}
\arguments{
\item{x}{dpp object}

\item{target}{name of dataset}

\item{...}{not used}
}
\value{
updated object
}
\description{
Add a age to an existing dpp object
}
\details{
Add a age to an existing dpp object
}
\examples{
# see help(dpp)
}
\seealso{
addComo, addSex, addVariable
}
\author{
Thomas A. Gerds <tag@biostat.ku.dk>
}
