% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dpp.R
\name{addBaseline<-}
\alias{addBaseline<-}
\title{Add baseline}
\usage{
addBaseline(x, name, ...) <- value
}
\arguments{
\item{x}{dpp object}

\item{name}{name of baseline date}

\item{...}{not (yet) used}

\item{value}{Instructions}
}
\description{
Add a baseline to a dpp object
}
\details{
Add a baseline to a dpp object
}
