% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dpp.R
\name{addComo}
\alias{addComo}
\title{Add comorbidity}
\usage{
addComo(x, target, como, ...)
}
\arguments{
\item{x}{dpp object}

\item{target}{name of dataset}

\item{como}{character string which should match \code{names(diseasecode)}}

\item{...}{not used}
}
\value{
updated object
}
\description{
Add a comorbidity to an existing dpp object
}
\details{
Add a comorbidity to an existing dpp object
}
\examples{
# see help(dpp)
}
\seealso{
addSex, addAge, addVariable, addDrug
}
\author{
Thomas A. Gerds <tag@biostat.ku.dk>
}
