% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dpp.R
\name{addData<-}
\alias{addData<-}
\title{Register a register}
\usage{
addData(x, name, id = "pnr", variables = NULL, ...) <- value
}
\arguments{
\item{x}{dpp object}

\item{name}{name of register. defaults to \code{"Register"}.}

\item{id}{variable to identify subjects across the registers. any other name than \code{"pnr"}
will internally be set to \code{"pnr"}. Note this will by side effect change the name
of id variable if the input is a \code{data.table}.}

\item{variables}{select variables from register.}

\item{...}{additional arguments passed to \code{addData} for
reading register from file.}

\item{value}{either a data.table or path to a file.}
}
\description{
Register a register
}
