% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dpp.R
\name{addDrug}
\alias{addDrug}
\title{Add drug}
\usage{
addDrug(x, target, drug, ...)
}
\arguments{
\item{x}{dpp object}

\item{target}{name of dataset}

\item{drug}{character string which should match \code{names(diseasecode)}}

\item{...}{not used}
}
\value{
updated object
}
\description{
Add a drug to an existing dpp object
}
\details{
Add a drug to an existing dpp object
}
\examples{
# see help(dpp)
}
\seealso{
addSex, addAge, addVariable, addComo
}
\author{
Thomas A. Gerds <tag@biostat.ku.dk>
}
