% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dpp.R
\name{addExclusion<-}
\alias{addExclusion<-}
\title{Inclusion instructions}
\usage{
addExclusion(x, name, target = "all", priority = c("late", "early"), ...) <- value
}
\arguments{
\item{x}{dpp object}

\item{name}{name of inclusion criterion}

\item{target}{currently ignored}

\item{priority}{Character. Data preprocessing time point. Either \code{"late"} or \code{"early"}.}

\item{...}{not (yet) used}

\item{value}{A function obtained with the all-purpose function \code{selector} or one of the specialized functions. See \code{dpp}.}
}
\description{
Instructions for a \code{dpp} object regarding which subjects to include
}
\details{
An exclusion criterion removes pnr numbers from one or several of the processed data sets.
}
