% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dpp.R
\name{addSex}
\alias{addSex}
\title{Add sex}
\usage{
addSex(x, target, ...)
}
\arguments{
\item{x}{dpp object}

\item{target}{name of dataset}

\item{...}{not used}
}
\value{
updated object
}
\description{
Add sex to an existing dpp object
}
\details{
Add sex to an existing dpp object
}
\examples{
# see help(dpp)
}
\seealso{
addComo, addAge, addVariable
}
\author{
Thomas A. Gerds <tag@biostat.ku.dk>
}
