% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addUsualRegisters.R
\name{addUsualRegisters}
\alias{addUsualRegisters}
\title{Add usual registers to dpp object}
\usage{
addUsualRegisters(x, pop = TRUE, dod = TRUE, lmdb = TRUE,
  lpr = TRUE, opr = TRUE, ...)
}
\arguments{
\item{x}{dpp object}

\item{pop}{If character this is the filename (full path) of the pop registry. If \code{TRUE} we search for pop.sas7bdat in result of \code{listRawdata(full.names=TRUE)}.}

\item{dod}{If character this is the filename (full path) of the dod registry. If \code{TRUE} we search for dod.sas7bdat in result of \code{listRawdata(full.names=TRUE)}.}

\item{lmdb}{If character this is the filename (full path) of the lmdb registry. If \code{TRUE} we search for lmdb.sas7bdat in result of \code{listRawdata(full.names=TRUE)}.}

\item{lpr}{If character this is the filename (full path) of the lpr registry. If \code{TRUE} we search for lpr.sas7bdat in result of \code{listRawdata(full.names=TRUE)}.}

\item{opr}{If character this is the filename (full path) of the opr registry. If \code{TRUE} we search for opr.sas7bdat in result of \code{listRawdata(full.names=TRUE)}.}

\item{...}{not yet used}
}
\value{
the updated object
}
\description{
Add usual registers to dpp object
}
\details{
Add usual registers to dpp object
}
\examples{
# data preprocessing object
\dontrun{
x <- dpp()
# locate registries
listRawdata(full.names=TRUE)
# add selected registries
x <- addUsualRegisters(x)
}
}
\seealso{
dpp
}
\author{
Thomas A. Gerds <tag@biostat.ku.dk>
}
