% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dpp.R
\name{addVariable<-}
\alias{addVariable<-}
\title{Add a variable to a dpp object}
\usage{
addVariable(
  x,
  target,
  data,
  type = c("merge", "join", "rbind"),
  depends = NULL,
  priority = c("late", "early"),
  by = NULL,
  ...
) <- value
}
\arguments{
\item{x}{dpp object}

\item{target}{Character. The name of a baseline date or \code{"fup"} for the followup dataset. Where to store the result.}

\item{data}{Optional the name of the data table in which to evaluate \code{value}.}

\item{type}{Character that determines how to add the variable to the data set. One of \code{"merge"}, \code{"join"} or \code{"rbind"}}

\item{depends}{vector of names of other instructions (inclusion, exclusion, variables) that should be evaluated before this one.}

\item{priority}{Processing priority. Maybe changed by \code{depends}.}

\item{by}{Character vector. Specifying subsets in which \code{value} is evaluated. E.g., if set to \code{"pnr"} then \code{value} is applied to the data of each subject separately.}

\item{...}{not (yet) used}

\item{value}{A function obtained with the all-purpose function \code{selector} or one of the specialized functions. See \code{dpp}.}
}
\description{
All purpose function to add a variable, more precisely
instructions (a function) which creates the variable (can use one
or all of the regis data)
}
