% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/admdb.R
\name{admdb<-}
\alias{admdb<-}
\title{Update database}
\usage{
admdb(dpp, id = pnr, inddto = inddto, uddto = uddto, add = FALSE) <-
  value
}
\arguments{
\item{dpp}{a}

\item{id}{a}

\item{inddto}{a}

\item{uddto}{a}

\item{add}{a}

\item{value}{a}
}
\description{
Update date database
}
\details{
Update database
}
\author{
Helene Charlotte Rytgaard
}

