% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/aggregateData.R
\name{aggregateData}
\alias{aggregateData}
\title{Aggregation of data - useful for poisson regression}
\usage{
aggregateData(data, timegrid)
}
\arguments{
\item{data}{dataset containing the variables "interval" and "event".}

\item{timegrid}{vector of times the data is aggregated by}
}
