% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drug.R
\name{atc}
\alias{atc}
\title{Add atc codes to object.}
\usage{
atc(atc)
}
\arguments{
\item{atc}{Quoted atc codes, a vector if multiple values.}
}
\examples{

d <- dpp()
drug(d, "treatment1") <- atc("A12B")

}
\author{
Helene Charlotte Rytgaard
}
