% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{b_codes}
\alias{b_codes}
\title{Danish treatment and care codes}
\format{A data table with 4676 rows and 2 variables
\describe{
   \item{kode}{A character vector of the code}
   \item{tekst}{A description of the treatment or care administered (in Danish)}
   }}
\source{
http://medinfo.dk/sks/
}
\usage{
b_codes
}
\description{
The Danish classification of treatment and care from SKS. Retrieved from SKS in October 2018
}
\keyword{datasets}
