% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cfh.R
\name{cfh}
\alias{cfh}
\title{Jannik's first R-version of the sas code from hell}
\usage{
cfh(diag)
}
\arguments{
\item{dt}{data.frame or data.table which provides variables: pnr, recno, inddto, uddto, pattype}
}
\value{
data.frame where all intermediate admission records are removed
}
\description{
R-version of the sas code from hell.
}
\details{
Ignores pattype
}
\examples{
data(samplepop)
cfh(samplepop)
}
\author{
Jannik Pallisgaard
}
