% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heaven-data.R
\docType{data}
\name{charlson.codes}
\alias{charlson.codes}
\title{charlson.codes}
\format{
A list
}
\usage{
charlson.codes
}
\description{
List with pre-specified codes for Charlson Index
}
\keyword{datasets}
