% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkMicroData.R
\name{checkMicroData}
\alias{checkMicroData}
\title{Check utable outputs for microdata}
\usage{
checkMicroData(dat, min.val = 5)
}
\arguments{
\item{dat}{Table to be checked. Must be the summarized output of Publish::utable()}

\item{min.val}{The minimum value allowed, anything below this value will be flagged as microdata}
}
\value{
A message showing the location of any microdata found, or a statement that no microdata was found
}
\description{
This function checks a summarized utable object for microdata. Currently it
checks all data colums and all rows. This means it may flag means, medians,
etc as microdata, even though they might be fine to export (i.e if the mean
value of some variable is 2, this function will flag it as microdata).
}
\examples{
checkMicroData(summary(Publish::utable(cyl ~ vs, data=mtcars)))
}
