% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/createProject.R
\name{createProject}
\alias{createProject}
\title{ordnung muss sein}
\usage{
createProject(directory)
}
\arguments{
\item{directory}{Full path of the new project. E.g., "v:/data/workdata/706818/BlaBlaBla/study"}
}
\description{
create new empty folder for project organisation
}
\details{
ordnung muss sein
}
\examples{
\dontrun{createProject("v:/data/workdata/706818/BlaBlaBla/study")}
}
\author{
Thomas A. Gerds <tag@biostat.ku.dk>
}
