% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drug.R
\name{drug<-}
\alias{drug<-}
\title{Add new drug}
\usage{
drug(d, drugname1) <- value
}
\arguments{
\item{dpp}{Data preprocessing object}

\item{add}{Logical variable. Per default set to FALSE. If TRUE, and drug with name already exists, then this is not overwritten.}

\item{name}{Name of the drug. (Just write, don't use "")}
}
\value{
The data preprocessing object is updated.
}
\description{
Update preprocessing object with new drug.
}
\author{
Helene Charlotte Rytgaard
}

