% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drugdb.R
\name{drugdb<-}
\alias{drugdb<-}
\title{Update database}
\usage{
drugdb(dpp, id = pnr, atc = atc, pdate = eksd, strength = strnum,
  npack = apk, ppp = packsize, add = FALSE) <- value
}
\arguments{
\item{dpp}{a}

\item{id}{a}

\item{atc}{a}

\item{pdate}{a}

\item{strength}{a}

\item{npack}{a}

\item{ppp}{a}

\item{add}{a}

\item{value}{a}
}
\description{
Update database
}
\details{
Update database
}
\author{
Helene Charlotte Rytgaard
}

