% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/findConditions.R
\name{findCondition}
\alias{findCondition}
\title{Find conditions in registry data}
\usage{
findCondition(data, id = "pnr", vars, keep, conditions,
  match = "contain", replace = NULL, combine = "long")
}
\arguments{
\item{data}{Data in which to search for conditions}

\item{vars}{Name(s) of variable(s) in which to search}

\item{keep}{a character vector of the columns in Data.table to keep in output}

\item{conditions}{A named list of (vectors of) search strings. See examples.}

\item{match}{A variable to tell how to use the character vectors: "exact"=exactly matches the search string, "contains"=contains the search string, "starts"=Starts with the search string, "end"=Ends with the search string}

\item{replace}{if replace=NULL then the set of variables corresponding to the names "conditions" will be inter 0/1 indicating whether the condition was met.  If replace is the name of a variable in Data, then that variable will be in the output replacing "1".}
}
\value{
A data table that includes the "keep-variables" and a set of ivariables corresponding to the names of the vectors in the list.  For each record these variables takes on values 0/1 when the condition is not-met/met OR NA/value-of-Replace-variable if replace is NULL/name-of-variable
}
\description{
Find conditions
}
\details{
Find conditions
}
\examples{
\dontrun{
library(data.table)
library(heaven)
d=simPrescriptionData(10000)
findCondition(data=d,vars="atc",conditions=list(drug1=c("N07","A12"),drug2=c("R03","R05")),match="start")
findCondition(data=d,vars="atc",conditions=list(drug1=c("N07","A12"),drug2=c("R03","R05")),match="start",combine="wide")
}
}
\author{
Christian Torp-Pedersen  <ctp@&hst.aau.dk>, Thomas A. Gerds <tag@biostat.ku.dk>
}
