% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heaven-data.R, R/heaven-package.R
\docType{data}
\name{Framingham}
\alias{Framingham}
\title{"Framingham"}
\format{A data.frame}
\usage{
Framingham
}
\description{
data.frame with Framingham data

The Framingham dataset is a public version of several examinations of people in the Framingham study. Various variables have been adjusted to ensure anonymity of people. Thus, the data cannot be used for science, only demonstration of R-code.
}
\examples{

data(Framingham)


}
\keyword{datasets}
