% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAdmLimits.R
\name{getAdmLimits}
\alias{getAdmLimits}
\title{Get Admission time}
\usage{
getAdmLimits(dt, pnr = "pnr", inddto = "inddto", uddto = "uddto")
}
\arguments{
\item{dt}{data.table containing the variables}

\item{pnr}{Variable with ID for each subject/group (default is \code{pnr}).}

\item{inddto}{Variable with entry times. Must be date or numeric (default is \code{inddto}).}

\item{uddto}{Variable with exit times. Must be date or numeric (default is \code{uddto}).}
}
\value{
List containing two datasets. 
        First contains the same data.table as for input but with 2 new variables first.inddto and last.outdate.
        Second contains all rows with missing data or where inddto > uddto.
}
\description{
Adds first admission date and last discharge date to a sequence of admissions. The default input data.table is the current standard
by which admissions are available in Statistics Denmark.
}
\examples{
\dontrun{
pop <- as.data.table(data(sample.diag))
## Variables have default names
newpop <- getAdmLimits(pop)

## Varaibles have custom names
setnames(pop, c('pnr','inddto','uddto'), c('personid', 'Entrydate','Outdate'))
newpop <- getAdmLimits(pop,pnr='personid',inddto='Entrydate',uddto='Outdate')

## List is returned containing two data.tables
newpopData <- newpop[[1]]
newpopError <- newpop[[2]]
}
}
\author{
Thomas A. Gerds <tag@biostat.ku.dk>
}
