% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAdmLimits2.R
\name{getAdmLimits2}
\alias{getAdmLimits2}
\title{Thomas's first R-version of the sas code from hell}
\usage{
getAdmLimits2(dt2)
}
\arguments{
\item{dt}{data.table which provides variables: pnr, recno, inddto, uddto, pattype}
}
\value{
nothing. but input data.table is modified such that all intermediate admission records are removed. This means that overlapping admission
intervals in the input data are combined into one admission.
}
\description{
R-version of the sas code from hell.
}
\details{
For now pattype is ignored.
}
\examples{
data(samplepop)
## take a copy to preserve the object samplepop 
pop=copy(samplepop)
setDT(pop)
newpop=getAdmLimits(pop)
print(print(newpop))
}
\author{
Thomas A. Gerds <tag@biostat.ku.dk>
}

