% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAdmLimits3.R
\name{getAdmLimits3}
\alias{getAdmLimits3}
\title{Get Admission time}
\usage{
getAdmLimits3(dt, personid, inddto, uddto)
}
\arguments{
\item{dt}{data.table}

\item{personid}{Variable with ID for each subject/group}

\item{inddto}{Variable with entry times. Must be date or integer}

\item{uddto}{Variable with exit times. Must be date or integer}
}
\value{
list containing dataset with admission dates add to data and dataset containing row errors
}
\description{
Removes row errors and returns data with add admission time
}
