% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAdmLimitsV2.R
\name{getAdmLimitsV2}
\alias{getAdmLimitsV2}
\title{Get Admission time}
\usage{
getAdmLimitsV2(dt, personid = "pnr", inddto = "inddto", uddto = "uddto")
}
\arguments{
\item{dt}{data.table containing the variables}

\item{personid}{Variable with ID for each subject/group. Default name pnr}

\item{inddto}{Variable with entry times. Must be date or numeric. Default name inddto}

\item{uddto}{Variable with exit times. Must be date or numeric. Default name uddto}
}
\value{
List containing dataset with admission dates add to data and dataset containing row errors
}
\description{
Removes row errors and returns data with add admission time
}
\examples{
\dontrun{
## Create Test data
Pnr <- c(1,1,2,2,2,3)
Entrydate <- as.Date(c("2001-03-14","2001-03-28","2006-03-14","2006-05-28","2001-11-11","1999-12-24"))
Outdate <- as.Date(c("2001-03-30","2001-04-09","2006-05-28","2006-06-21","2001-12-04","2000-02-13"))
Extra <- c(1:6)

library(data.table)
mydata <- data.table(Pnr,Entrydate,Outdate,Extra)

pop <- getAdmLimits3(mydata,personid='Pnr',inddto='Entrydate',uddto='Outdate')

## List is returned containing two data.tables
popData <- pop[[1]]
popError <- pop[[2]]
}
}
\author{
Thomas A. Gerds <tag@biostat.ku.dk>, Peter Enemark Lund <peter.l@rn.dk>
}
