% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dpp.R
\name{getRawdata}
\alias{getRawdata}
\title{read raw data}
\usage{
getRawdata(x, id, n, ...)
}
\arguments{
\item{x}{dpp object}

\item{id}{id variable}

\item{n}{sample size}

\item{...}{passed to subroutine: fread or importSAS}
}
\description{
read raw data
}
\details{
read raw data
}
\author{
Thomas A. Gerds <tag@biostat.ku.dk>
}
