% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/groupAges.R
\name{groupAges}
\alias{groupAges}
\title{groupAges}
\usage{
groupAges(
  x,
  lower,
  upper,
  by,
  sep = " - ",
  include.below = TRUE,
  include.above = TRUE,
  below.char = "<",
  above.char = "+"
)
}
\arguments{
\item{x}{Vector to be grouped (usually ages in years)}

\item{lower}{Lower bound (inclusive) of age}

\item{upper}{Upper bound (exclusive) of age}

\item{by}{How many years to group by}

\item{sep}{Controls the serperator in the label between the the lower and
upper bound for each age group}

\item{include.below}{If TRUE (default) value under the lower bound will be
included and labeled accordingly. If false, values under will be returned as
NA}

\item{include.above}{If TRUE (default) value over the upper bound will be
included and labeled accordingly. If false, values over will be returned as
NA}

\item{below.char}{The character to use for labeling the group of value under
the lower bound}

\item{above.char}{The character to use for labeling the group of value above
the lower bound}
}
\value{
Returns a factor vector of the same length as x.
}
\description{
Helper function to bin a vector of continuous ages into groups
 and automatically generate labels.
}
\examples{
library(heaven)
set.seed(1)
x <- floor(runif(20, min = 0, max = 105))
groupAges(
 x,
 lower = 15,
 upper = 50,
 by = 10,
 below.char = "<",
 above.char = "+"
)
}
\author{
Matthew Phelps
}
