% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/halal.R
\name{halal}
\alias{halal}
\title{Extraction of diseases by diagnoses}
\usage{
halal(dat, disease = NULL, inclusions = NULL, exclusions = NULL,
  p.in = NULL, p.out = NULL, first.pnr = FALSE, pat = NULL,
  prefix = "", entryvar = "inddto", id = "pnr", outvar = "uddto",
  codevar = "diag", patvar = "pattype", record.id = "recnum")
}
\arguments{
\item{dat}{Raw data with diagnoses, needs specific variables: diag, pnr, inddto, and pattype}

\item{disease}{pre-specified name of disease after HALAL definitions, as characterstring. Can be one or more of: "af",}

\item{inclusions}{Characterstring, where other diagnoses can be included.}

\item{exclusions}{Characterstring, specifying diagnoses to be omitted - if desired.}

\item{p.in}{Date of period start}

\item{p.out}{Date of period end}

\item{first.pnr}{Specifies if only the first record of each patient should be output}

\item{pat}{Number or vector defining types of patients to include (pattype: 0,1,2,3), default is all types.}

\item{prefix}{character string of prefix name for the resulting date variable of disease.}
}
\author{
Regitze Kuhr Skals
}

