% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heaven-data.R
\docType{data}
\name{hypertensionATC}
\alias{hypertensionATC}
\title{hypertensionATC}
\format{
A list
}
\usage{
hypertensionATC
}
\description{
List with pre-specified ATC codes for classes of antihypertensive medication
}
\keyword{datasets}
