% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dpp.R
\name{importer}
\alias{importer}
\title{import data from a register}
\usage{
importer(data, residents = TRUE, where, keep = "pnr", select = "unique", ...)
}
\arguments{
\item{data}{names of data}

\item{residents}{only residents?}

\item{where}{SAS where statement}

\item{keep}{which variables to keep}

\item{select}{select subset. currently only \code{"unique"} is supported}

\item{...}{passed to importSAS}
}
\value{
data.table
}
\description{
import data from a register
}
\details{
import data from a register
}
\seealso{
dpp
}
\author{
Thomas A. Gerds <tag@biostat.ku.dk>
}
