% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{innerMedicinMacro}
\alias{innerMedicinMacro}
\title{The heart of the medicin macro}
\usage{
innerMedicinMacro(dat, admdat, doses, idunique, prescriptionwindow,
  maxdepot)
}
\arguments{
\item{dat}{data set}

\item{admdat}{admission data}

\item{doses}{doses}

\item{idunique}{unique subject ids}

\item{prescriptionwindow}{prescription window}

\item{maxdepot}{see medicine macro}
}
\description{
Inner process of medicin macro
}
\author{
Helene Charlotte Rytgaard and Thomas Alexander Gerds
}
