% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{innerprocess}
\alias{innerprocess}
\title{The heart of the medicin macro}
\usage{
innerprocess(dat, admdat, doses, idunique, N, maxdepot, collapse)
}
\arguments{
\item{dat}{data set}

\item{admdat}{admission data}

\item{doses}{doses}

\item{idunique}{unique subject ids}

\item{N}{sample size}

\item{maxdepot}{see medicine macro}

\item{collapse}{If \code{TRUE} collapse admission periods when there is not a single day out of hospital in between.}
}
\description{
Inner process of medicin macro
}
\author{
Helene Charlotte Rytgaard
}
