% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internalFormat.R
\name{internalFormat}
\alias{internalFormat}
\title{internalFormat}
\usage{
internalFormat(dt, update.names = NULL, update.classes = NULL,
  autoformat = TRUE, silent = FALSE)
}
\arguments{
\item{dt}{\code{data.table} to format or check format for.}

\item{update.names}{Named list of names to updates.}

\item{update.classes}{Named list of variables to update (change the \code{class} for these variables).}

\item{autoformat}{If \code{TRUE} the function will try to format \code{dt}. Otherwise it will just check the formats to see if they match the ones specified in in \code{update.classes}. The names are always updated if \code{update.names} is non-empty.}

\item{silent}{If \code{TRUE} the function will print a message if there was any changes or problems with the formatting.}
}
\value{
The output is a list consisting of: (1) The (formatted) \code{data.table}, (2) a vector of strings denoting keywords about which kind of formatting problems (if any) occurred, and (3) the whole detailed messages about which variable(s) were problematic (if any).
}
\description{
Internal function for formatting dppObjects.
}
\details{
Internal function used in, e.g., format.dppObject.
It takes a data.table as input a tries to format it and/or checks the format.

The output is the (formated) \code{data.table} and a to string of information: One stating which kind of formatting problems (if any) was encountered, and a detailed description of the problems.

Note: The names are first updated according to \code{update.names} and then the classes are updated according to \code{update.classes}; thus \code{update.classes} should contain the updated names.
}
\author{
Anders Munch
}
