% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/internalLoadData.R
\name{internalLoadData}
\alias{internalLoadData}
\title{internalLoadDate}
\usage{
internalLoadData(dppObject)
}
\arguments{
\item{dppObject}{dppObject in the status "loaded" or "nominal".}
}
\value{
The output is a list consisting of: (1) The loaded and formatted \code{dppObject}, (2) a vector of strings denoting keywords about which kind of formatting problems (if any) occurred, and (3) the whole detailed messages about which variable(s) were problematic (if any).
##' @author Anders Munch
}
\description{
Internal function to load data into dpp.
}
\details{
Internal function used to load data into dpp.

The function checks if the dppObject is already loaded or only specified nominally (i.e., as a path to the data). If the latter is the case, the data is loaded; in any case the data is formatted, or at least tried to be formatted correctly according to the information in the dppObject.
The output is then the formatted dppObject and information about possible format or name conflicts.
}
