% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{k_codes}
\alias{k_codes}
\title{Danish operaction/surgery codes}
\format{A data table with 10270 rows and 2 variables
\describe{
   \item{kode}{A character vector of the code}
   \item{tekst}{A description of the operation/surgery (in Danish)}
   }}
\source{
http://medinfo.dk/sks/
}
\usage{
k_codes
}
\description{
The Danish classification system for operations/surgeries as used in SKS. Retrieved from SKS in October 2018
}
\keyword{datasets}
