% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heaven-data.R
\docType{data}
\name{kommune_codes}
\alias{kommune_codes}
\title{kommune_codes}
\format{
A data.table
}
\usage{
kommune_codes
}
\description{
data.table with the names and codes of every kommune, and the region they belong to
}
\keyword{datasets}
