% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kommuneRegion.R
\name{kommuneRegion}
\alias{kommuneRegion}
\title{kommuneRegion - Converts kommune number to Region}
\usage{
kommuneRegion(kommune)
}
\arguments{
\item{kommune}{3-digit kommune number.}
}
\description{
Converts a 3-digit kommune number to text representing one of the five
administrative regions.
}
\details{
This function converts a 3-digit kommune number to text corresponding to
one of the five regions in Denmark. The conversion is adapted from a DST
SAS format and is valid from 2007. There is no easy way to convert older
kommune numbers to the current regions, since some municipalities were split
in the process of creating the five regions.
}
\examples{
kommuneRegion(410) # Sydjylland

}
\author{
Christian Torp-Pedersen
}
