% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lexisFromTo.R
\name{lexisFromTo}
\alias{lexisFromTo}
\title{lexisFromTo}
\usage{
lexisFromTo(indat,splitdat,invars,splitvars)
}
\arguments{
\item{indat}{- base data with id, start, end, event and other data - possibly 
already split}

\item{splitdat}{- Data with splitting guide - id/from/to/value/name}

\item{invars}{- vector of colum names for id/entry/exit/event - in that 
order, 
example: c("id","start","end","event")}

\item{splitvars}{- vector of column names containing dates to split by.
example: c("id","start","end","value","name") - must be in that order!}
}
\value{
The function returns a new data table where records have been split according 
to the splittingguide dataset. Variables
unrelated to the splitting are left unchanged.
}
\description{
LexisFromTo is one of 3 lexis functions in heaven.  The purpose is to split 
observations according to possibly temporary changes. Typical situations
are drug therapy, pregnancy, a period after surgery etc. The input is two
data.tables.  The "base" data are the data to be split. They may contain
much information, but the key is "id","start","end" and "event". These des-
cribe the participant's id, start of time period, end of time period and the
event of interest (must be 0/1).

The other dataset is the "splittingguide". This may include information to 
guide splitting from multiple sources (multiple difference medications, se-
veral pregnancies etc). This should have the following variables: "id",
"start","end","value","name". It is the participant id, start/end of next
interval, a "value" (character, dose of drug, presence of pregnancy etc), and
finally a "name" which defines the column in the output data with information
from that particular split.
}
\details{
The input to this function are two data.tables and two lists of the critical 
variables.  The BASE data it the data to be split.
This data must have a variable to identify participants, start/end times and 
a variable to indicate event after last interval.
The other table (SPLITTINGUIDE) contains possibly multiple records for each 
participants with id/from/to/value/name.
The program checks that intervals are not negative and that intervals for 
each "name" and each individual do not overlap.  Violation stops the program 
with error. Overlaps may occur in real situations, but the user needs to make 
decisions regarding this prior to using this function.

It is required that the splittingguide contains at least one record.  
Missing data for key variables are not allowed and will cause errors.

A note of caution: This function works with dates as integers. R has a de-
fault origina of dates as 1 January 1970, but other programs have different
default origins - and this includes SAS and Excell. It is therefor important
for decent results that care is taken that all dates are defined similarly.

The output will always have the "next" period starting on the day where the
last period ended. This is to ensure that period lengths are calculated pro-
perly. The program will also allow periods of zero lengths which is a conse-
quence when multiple splits are made on the same day. When there is an event
on a period with zero length it is important to keep that period not to 
loose events for calculations. Whether other zero length records should be
kept in calculations depend on context.
}
\examples{
library(data.table)
dat <- data.table(id=c("A","A","B","B"),
                 start=c(0,100,0,100),
                 end=c(100,200,100,200),
                 event=c(0,1,0,0))
split <- data.table (id=c("A","A","A","B","B","B"),
                    start=c(0,50,25,110,150,400),
                    end= c(25,75,150,120,250,500),
                    value=c(1,2,3,1,2,3),
                    name=c("d1","d1","d2","d1","d1","d2"))
temp <- lexisFromTo(dat # inddato with id/in/out/event
                   ,split # Data with id and dates
                   ,c("id","start","end","event") #names of id/in/out/event - in that order
                   ,c("id","start","end","value","name")) #Nmes var date-vars to split by
temp
}
\author{
Christian Torp-Pedersen
}
