% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lexisTwo.R
\name{lexisTwo}
\alias{lexisTwo}
\title{lexisTwo}
\usage{
lexisTwo(indat,splitdat,invars,splitvars)
}
\arguments{
\item{indat}{A data.table or data.frame whose first 4 columns are in that order:
\item{id}{person identification variable such as PNR. The data may contain multiple lines per subject.}
\item{start}{Start of time period. Either in date format or given as numeric/integer.}
\item{end}{End of time period. Either in date format or given as numeric/integer.}
\item{event}{Event}}

\item{splitdat}{- Data with splitting guide - id and columns with dates to 
split by}

\item{invars}{- vector of colum names for id/entry/exit/event - in that 
order, example: c("id","start","end","event")}

\item{splitvars}{- vector of column names of columns containing dates to 
split by. example: c("date1","date2","date3","date4")
The name of the id column must be the same in both datasets}
}
\value{
The function returns a new data table where records have been split according 
to the splittingguide dataset. Variables unrelated to the splitting are 
left unchanged. The names of columns from "splitvars" are also in output
data, but now they have the value zero before the dates and 1 after.
}
\description{
lexixTwo is one of 3 splitting functions in heaven.
This particular function (lexixTwo) is useful to add
the time-dynamic occurrence of comorbidities and other events
to an existing data set which contains person specific information
about periods in time. The person specific time periods are split 
according to the occurrence dates of the comorbidities and other events.

The "base" data are the data to be split. They 
may contain much information, but the key is "id","start","end" and "event". 
These describe the participant's id, start of time period, end of time period 
and the event of interest (must be 0/1).

The other input is a data.table with the splittingguide. This data.table 
should have one record pr. individual.  One column defined the same id as in
the "base" table. The other columns contain dates for each condition where 
the split should occur.  These column names will also appear in the output
data, but on output the values are zero before the dates and 1 after.  When
dates are NA output has zero.
}
\details{
#' The program checks that intervals are not negative. Violation stops  
with error. Overlaps may occur in real situations, but the user needs to make 
decisions regarding this prior to using this function.

It is required that the splittingguide contains at least one record.  
Missing data for key variables are not allowed and will cause errors.

A note of caution: This function works with dates as integers. R has a de-
fault origina of dates as 1 January 1970, but other programs have different
default origins - and this includes SAS and Excell. It is therefor important
for decent results that care is taken that all dates are defined similarly.

The output will always have the "next" period starting on the day where the
last period ended. This is to ensure that period lengths are calculated pro-
perly. The program will also allow periods of zero lengths which is a conse-
quence when multiple splits are made on the same day. When there is an event
on a period with zero length it is important to keep that period not to 
loose events for calculations. Whether other zero length records should be
kept in calculations depend on context.
}
\examples{
library(data.table)
dat <- data.table(id=c("A","A","B","B","C","C","D","D"),
                start=c(0,100,0,100,0,100,0,100),
                end=c(100,200,100,200,100,200,100,200),
                event=c(0,1,0,0,0,1,0,1))
split <- data.table (id=c("A","B","C","D"),date1=c(0,50,150,300),
                 date2=c(25,75,175,325),date3=c(30,30,30,30),
                    date4=c(0,1,0,1))
temp <- lexisTwo(dat # inddato with id/in/out/event
       ,split # Data with id and dates
       ,c("id","start","end","event") #names of id/in/out/event - in that order
       ,c("date1","date2","date3","date4")) #Names var date-vars to split by
temp
}
\seealso{
lexisSeq lexisFromTo
}
\author{
Christian Torp-Pedersen
}
