% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/listRawdata.R
\name{listRawdata}
\alias{listRawdata}
\title{List sas data files}
\usage{
listRawdata(dir, pattern = "\\\\.sas7bdat$", full.names = FALSE, ...)
}
\arguments{
\item{dir}{directory which contains the sas files}

\item{pattern}{passed to \code{list.files}}

\item{full.names}{passed to \code{list.files}}

\item{...}{passed to list.files}
}
\value{
list with file names
}
\description{
List the contents of a directory which contains sas data files
}
\details{
List the contents of a directory which contains sas data files
}
\examples{
\dontrun{listRawdata("x:/Data/Rawdata_Hurtig/706818/")}
}
\seealso{
\code{scanRawdata}, \code{addUsualRegisters}
}
\author{
Thomas A. Gerds <tag@biostat.ku.dk>
}
