% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RcppExports.R
\name{Matcher}
\alias{Matcher}
\title{Matcher}
\usage{
Matcher(Ncontrols, Tcontrols, Ncases, reuseControls, controlIndex, caseIndex,
  controls, cases, NoIndex)
}
\description{
Defines controls for cases
}
\author{
Christian Torp-Pedersen
}
