% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxdepot.R
\name{maxdepot<-}
\alias{maxdepot<-}
\title{Update database}
\usage{
maxdepot(dpp, ...) <- value
}
\arguments{
\item{dpp}{a}

\item{...}{a}

\item{value}{a}
}
\description{
maximum repositary
}
\details{
maximum repositary
}
\author{
Helene Charlotte Rytgaard
}

