% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/maxrep.R
\name{maxrep<-}
\alias{maxrep<-}
\title{Update database}
\usage{
maxrep(dpp, ...) <- value
}
\arguments{
\item{dpp}{a}

\item{...}{a}

\item{value}{a}
}
\description{
maximum repositary
}
\details{
maximum repositary
}
\author{
Helene Charlotte Rytgaard
}

