% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medicinMacro.R
\name{medicinMacro}
\alias{medicinMacro}
\title{Medicin macro to estimate prescription lengths and averages}
\usage{
medicinMacro(
  drugs,
  drugdb,
  admdb,
  datesdb,
  type = "dynamic",
  drugdb.datevar = "eksd",
  admdb.datevars = c("first.indate", "last.outdate"),
  datesdb.datevars,
  method,
  cap.values,
  window,
  pre.window,
  id = "pnr",
  strength.var = "strnum",
  packsize.var = "packsize",
  atc.var = "atc",
  apk.var = "apk",
  splitting = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{drugs}{A named list of drugs. Each element of this list should be a list
with the following elements:
\describe{
  \item{atc}{A vector of ATC codes which should match the components of the drug exactly.}
  \item{maxdepot}{The maximum total dose that a single patient can possibly stack.}
  \item{period}{A vector of dates to limit the period in which to estimate the daily dose.}
  \item{prescriptionwindow}{Default is 2 prescriptions.}
  \item{doses}{A named list with the elements \code{value}, \code{min}, \code{max} and \code{def}.
  Here \code{value} is a vector of strengths of one unit (e.g. pill) of the drug.
  The vector should have one such strength for each of the different packages that occur in the data.
  \code{min} is a vector of the same length as \code{value} where each element is the assumed minimum total dose that a patient can consume on one day.
  \code{max} is a vector of the same length as \code{value} where each element is the assumed maximum total dose that a patient can consume on one day.
  \code{def} is a vector of the same length as \code{value} where each element is the assumed default dose that an average patient would consume on one day.
  See examples.}
}}

\item{drugdb}{data.table with (subset of) medical drugs registry}

\item{admdb}{data.table with (subset of) hospital admission registry. The data.table should be
prepared such that it contains only overnight hospital stay (i.e., pattype=2) and non-overlapping
hospital stay periods, i.e., as obtained with \code{getAdmLimits} (SAS-AKA: code-from-hell).}

\item{datesdb}{data.table with individual periods or time points of interest.
Only used when the \code{type} argument is different from \code{"dynamic"}.
Should contain one start and end date for each individual when \code{type} equal to \code{"period"}
and just one date for each individual when \code{type} equal to \code{"cross-sectional"}
(see examples).}

\item{type}{string, specifying which analysis to perform.
Can be \code{"dynamic"}, \code{"period"}, or \code{"cross-sectional"}. Default is \code{"dynamic"}.
(see examples).}

\item{drugdb.datevar}{name of the date variable in \code{drugdb}. Default is \code{"eksd"}.}

\item{admdb.datevars}{names of the date variables in \code{admdb}. Default is \code{c("inddto","uddto")}.}

\item{datesdb.datevars}{name(s) of the date variables in \code{datesdb}.
Should be a single string when \code{type}=\code{"cross-sectional"} and a vector of two string when
\code{type}=\code{"period"}. Default is \code{"time"} in the first case and \code{c("start", "end")} in the second.}

\item{method}{Which method to use to estimate the daily dose. Can be \code{"default.doses"} or
\code{"number.of.days"}. Default is \code{"default.doses"}.
See documentation for more details.
Only used when the \code{type} is different from \code{"dynamic"}.}

\item{cap.values}{Boolean, specifying whether or not to restrict the estimated daily dose to lie
within the specified minimum and maximum dose. Default is \code{TRUE}
Only used when the \code{type} is different from \code{"dynamic"}.}

\item{window}{Only used when the \code{type} = \code{"cross-sectional"}.
Specify how many days back from the given time point we should go to find purchases to estimate the daily dose.}

\item{pre.window}{Only used when the \code{type} = \code{"period"}.
Specify how many days back from the individual end periods we should go to find purchases to estimate the
daily dose for the first part of the periods.}

\item{id}{Name of variable to identify individuals. Default is
\code{"pnr"}.}

\item{strength.var}{Name of variable to identify strength. Default
is \code{"strnum"}.}

\item{packsize.var}{Name of variable to identify size of
packages. Default is \code{"packsize"}.}

\item{atc.var}{Name of variable to identify drugs. Default is \code{"atc"}.}

\item{apk.var}{Name of variable to identify number of
packages. Default is \code{"apk"}.}

\item{splitting}{Split the data into 10 chunks to estimate
remaining time (and possibly speed up computation).}

\item{verbose}{bla bla?}
}
\description{
Processing function, to perform calculations according to the data and variables specified in preprocessing object.
}
\author{
Helene Charlotte Rytgaard, Anders Munch and Thomas Alexander Gerds based
    on the famous SAS macro by Christian Tobias Torp-Pedersen
}
