% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/N.R
\name{N<-}
\alias{N<-}
\title{Update database}
\usage{
N(dpp, ...) <- value
}
\arguments{
\item{dpp}{a}

\item{...}{a}

\item{value}{a}
}
\description{
Update database
}
\details{
Update database
}
\author{
Helene Charlotte Rytgaard
}

