% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/nccSampling.R
\name{nccSampling}
\alias{nccSampling}
\title{Dataset for nested case-control}
\usage{
nccSampling(pnr, time, status, Ncontrols = 10L, data = NULL,
  match = NULL, include = NULL, Tstart = rep(0, length(pnr)),
  exposureWindow = 0, cores = 1)
}
\arguments{
\item{pnr}{PNR for the individual patients}

\item{time}{Time of event or censoring}

\item{status}{Event indicator}

\item{Ncontrols}{Number of controls per case}

\item{data}{Dateset containing the variables (not the matching and inclusion variables yet)}

\item{match}{List of matching variables}

\item{include}{List of covariates to include in the output dataset}

\item{Tstart}{Time of treatment start}

\item{exposureWindow}{Minimum required time of exposure to be included both as a case and as a control}

\item{cores}{Number of cores to use}
}
\value{
A data.table with the variables pnr, time, status and chosen covariates
}
\description{
Dataset for nested case-control
}
\details{
The time variable can be used as strata variable in clogit
}
\examples{
\dontrun{n <- 100000
pnr <- 1:n
time <- rexp(n)
status <- sample(0:1,n,replace=TRUE,prob=c(.99,.01))
sex <- sample(c("M","F"),n,replace=TRUE)
age <- sample(c("y","m","o"),n,replace=TRUE)
d <- nccSampling(pnr,time,status,Ncontrols=5,match=list(sex=sex,age=age))
}
}
\author{
Jeppe E. H. Madsen <vcl891@alumni.ku.dk>
}
