% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/new_dppObject.R
\name{new_dppObject}
\alias{new_dppObject}
\title{new_dppObject}
\usage{
new_dppObject(formats)
}
\arguments{
\item{formats}{Specify the formats of the dppObject. Should be a named list of formats.}
}
\value{
An object of class \code{dppObject} with specified formats but otherwise empty.
}
\description{
Internal function to setup new dppObject.
}
\details{
Internal function to create dppObject with some structure. It is only used through other functions.
}
\author{
Anders Munch
}
