% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/period.R
\name{period<-}
\alias{period<-}
\title{Update database}
\usage{
period(dpp) <- value
}
\arguments{
\item{dpp}{a}

\item{value}{a}
}
\description{
Update database
}
\details{
Update database
}
\author{
Helene Charlotte Rytgaard
}

