% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotoutput.R
\name{plot.dppout}
\alias{plot.dppout}
\title{Update database}
\usage{
\method{plot}{dppout}(out, drug = NULL, id = NULL, which = 2)
}
\arguments{
\item{id}{a}

\item{dpp}{a}

\item{trace}{a}
}
\description{
Function to plot output etc
}
\details{
Function to plot output etc
}
\author{
Helene Charlotte Rytgaard
}

