% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poissonregression.R
\name{poissonregression}
\alias{poissonregression}
\title{Poisson regression by aggregated data using glm}
\usage{
poissonregression(formula, data, timegrid, effectZgrid)
}
\arguments{
\item{formula}{The formula as specified in glm.}

\item{data}{dataset containing the variables "interval" and "event".}

\item{timegrid}{vector of times the data should be aggregated by.}
}
