% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/poissonRisk.R
\name{poissonRisk}
\alias{poissonRisk}
\title{Predicting absolute risk from poisson regression model}
\usage{
poissonRisk(model, interval, newdata)
}
\arguments{
\item{model}{glm object of fitted poisson model}

\item{interval}{risk interval within time scale of the poisson model}

\item{newdata}{values of the covariates in the poisson model for which the absolute risks are calculated.}
}
\author{
Thomas A. Gerds <tag@biostat.ku.dk>, Regitze Kuhr Skals <r.skals@rn.dk>
}
