% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/preprocess.R
\name{preprocess}
\alias{preprocess}
\title{Update database}
\usage{
preprocess(dpp, id = NULL, trace = FALSE)
}
\arguments{
\item{dpp}{a}

\item{id}{a}

\item{trace}{a}
}
\description{
Update database
}
\details{
Update database
}
\author{
Helene Charlotte Rytgaard
}

