% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{process}
\alias{process}
\title{Update database}
\usage{
process(dpp, treatments = NULL, id = NULL, trace = FALSE, out = TRUE)
}
\arguments{
\item{dpp}{a}

\item{id}{a}

\item{trace}{a}
}
\description{
Update database
}
\details{
Update database
}
\author{
Helene Charlotte Rytgaard
}

