% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dpp.R
\name{read.register}
\alias{read.register}
\title{Read a registered register from SAS data file or csv file}
\usage{
read.register(file, id, variables, ...)
}
\arguments{
\item{file}{file name}

\item{id}{Character. Subject id variable. Default is \code{"pnr"}.}

\item{variables}{Read only these variables.}

\item{...}{Arguments passed to \code{importSAS} or \code{fread}.}
}
\value{
data.table
}
\description{
Read a register
}
\details{
Read a register for data preprocessing purposes.
}
\seealso{
dpp
}
\author{
Thomas A. Gerds <tag@biostat.ku.dk>
}
