% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heaven-package.R
\docType{data}
\name{sampleDrugData}
\alias{sampleDrugData}
\title{Sample drug purchase data for illustration}
\format{This data frame contains 12 observations from one subject:
\describe{
\item{pnr}{pnr number}
\item{eksd}{date of purchase}
\item{packsize}{package size}
\item{strnum}{strnum}
\item{apk}{apk}
\item{atc}{atc code}
}}
\description{
a sample data frame used to illustrate some functionality
}
\examples{

data(sampleDrugData)


}
\keyword{datasets}
