% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heaven-package.R
\docType{data}
\name{samplepop}
\alias{samplepop}
\title{samplepop}
\format{A data frame with 32 observations on the following 5 variables.
\describe{
\item{pnr}{patient id}
\item{recno}{record number}
\item{inddto}{hospital admission first date}
\item{uddto}{hospital admission last date}
\item{pattype}{type of admission}
}}
\description{
sample data to test R versions of code from hell
}
\examples{

data(samplepop)

}
\keyword{datasets}

