% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/scanRawdata.R
\name{scanRawdata}
\alias{scanRawdata}
\title{Scan sas data files}
\usage{
scanRawdata(
  dir,
  which = c("pop", "doede", "diag", "lmdb", "opr"),
  verbose = FALSE
)
}
\arguments{
\item{dir}{directory which contains the sas files}

\item{which}{vector of expressions to match the file names}

\item{verbose}{passed to importSAS}
}
\value{
list of file contents
}
\description{
Scan the contents of a directory which contains sas data files
}
\details{
Scan the contents of a directory which contains sas data files
}
\examples{
\dontrun{
scanRawdata("x:/Data/Rawdata_Hurtig/706818/",which=c("pop","lmdb"))
}
}
\seealso{
\code{listRawdata}
}
\author{
Thomas A. Gerds <tag@biostat.ku.dk>
}
