% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/searchEvent.R
\name{searchEvent}
\alias{searchEvent}
\title{Search for an event}
\usage{
searchEvent(dt, periods, atcs = NULL, interval = NULL, pnr = "pnr",
  atc = "atc", date = "date", startdate = "startdate",
  enddate = "enddate")
}
\arguments{
\item{dt}{data.table to search through for event. Should at least contain an id ('pnr') and a date ('date').}

\item{periods}{Periods in which to search for an event. Can be either a single date or a data.table with individual dates. See details.}

\item{atcs}{Vector of atc codes. If this is supplied, the detected event will also be requied to have one of the given atc codes (see details).}

\item{interval}{Number of day to look back from the start dates. See details.}

\item{pnr}{Variable name to identify the individual. Should be the same in both 'dt' and 'periods' (if this is a data.table).}

\item{atc}{Variable name to identify the atc codes in 'dt'.}

\item{date}{Variable name to identify the date for an event in 'dt'.}

\item{startdate}{Variable name to identify the start date for the event search in 'periods'.}

\item{enddate}{Variable name to identify the end date for the event search in 'periods'.}
}
\value{
data.table of unique individuals and whether or not an event happened for each individual. See details.
}
\description{
Search for an event in specified time period for each individual.
}
\details{
When supplying a data.table in 'periods' this should as minimum include an id variable (with the same name as the id variable in 'dt') and a start date. If an end date is also supplied, the search for an event will be be in the interval from end data to start date (so end date should always be before start date). If no end date is supplied, the periods to search in will be determined by the argument 'interval' by searching in the period from start data - interval to start date; so every person can have a unique start date, but to period to search through has the same length for all persons. ##' 
If only a single date is given in 'periods' the periods are again based on 'interval' and will be the same periods for all person: From start date - interval to start date.

If an argument to 'atcs' is supplied the event search is carried out for each specified atc code. The output will be a data.table with unique individuals and one column for each atc code, specifying whether or not the given act code was detected in the period for each individual.
If no argument is given, the output will simply be a data.table with unique individuals and a column denoting whether or not an event was detected in the given (individual) period.
}
\examples{

set.seed(333)
n <- 200
dt <- simPrescriptionData(n, startDate="2005-01-01")

## Look at the data
dt

## See how many individuals purchased something with 2006-06-01 and 100 days back.
searchEvent(dt=dt,periods=as.Date("2006-06-01"),interval=100, date="eksd")

## Searching with indivual periods for subsample of individuals
## Simulate some dates
periods <- data.table::copy(dt[1:101,])
periods$pnr <- sample(1:n,101)
periods$eksd <- periods$eksd+round(rnorm(101,10,sd=10))
periods$enddate <- periods$eksd-sample(1:400,size=101,replace=TRUE)

## Look at the data
periods

## Do the search, now with individual dates
searchEvent(dt=dt,periods=periods,date="eksd",startdate="eksd")

## Specifying the atcs will look for these particular purchases
searchEvent(dt=dt,periods=periods,atcs=c("A07", "A12B"),date="eksd",startdate="eksd")
 
}
\author{
Anders Munch
}
