% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dpp.R
\name{selector}
\alias{selector}
\title{Selecting and collecting information from raw data}
\usage{
selector(
  data,
  var,
  name,
  varnames,
  by = "pnr",
  search.term = NULL,
  search.expression = NULL,
  sortkey = NULL,
  period = NULL,
  backward = NULL,
  forward = NULL,
  select = "first",
  collect = NULL,
  ...
)
}
\arguments{
\item{data}{Name of raw data source}

\item{var}{raw data variables}

\item{name}{name of this action}

\item{varnames}{names for processed variable}

\item{by}{by variable}

\item{search.term}{search term applied to var}

\item{search.expression}{search expression applied to var}

\item{sortkey}{in this order}

\item{period}{list with three elements: variable, start, stop}

\item{backward}{search backward in time}

\item{forward}{search forward in time}

\item{select}{what to select}

\item{collect}{what to collect}

\item{...}{additional arguments}
}
\description{
Selecting and collecting information from raw data to processed data
}
\details{
This is the master function for several different data processing tasks
}
\author{
Thomas Alexander Gerds tag@biostat.ku.dk
}
