% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simAdmissionData.R
\name{simAdmissionData}
\alias{simAdmissionData}
\title{Admission data simulation function}
\usage{
simAdmissionData(n, m = 5, diagnoses = paste0("D", toupper(letters),
  rep(0:99, length(letters))), startDate = "1995-01-01")
}
\arguments{
\item{n}{Number of patients}

\item{m}{Maximal number of admission dates per patient}

\item{diagnoses}{List of diagnoses. Defaults to all possible ICD10 codes: http://www.icd10data.com/ICD10CM/Codes}

\item{startDate}{starting date}
}
\description{
Simulate admission data
}
\details{
Simulate admission data alike the Danish medical registry
}
\examples{
## a single subject
set.seed(1)
simAdmissionData(1)
## 7 subjects
set.seed(28)
simAdmissionData(7)

}
\author{
Helene Charlotte Rytgaard \email{hely@sund.ku.dk}, Thomas Alexander Gerds \email{tag@biostat.ku.dk}
}
