% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simdata.R
\name{simdata}
\alias{simdata}
\title{Update database}
\usage{
simdata(which = 1, dos = c(75, 100, 50, 125))
}
\arguments{
\item{dpp}{a}

\item{id}{a}

\item{maxdepot}{a}

\item{trace}{a}
}
\description{
Update database
}
\details{
Update database
}
\author{
Helene Charlotte Rytgaard
}

