% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simpoisson.R
\name{simpoisson}
\alias{simpoisson}
\title{Simulate data for poisson regression}
\usage{
simpoisson(
  N,
  changepoints,
  baseline,
  beta,
  gamma,
  delta = 0,
  comprisk,
  alpha,
  zeta,
  px,
  pz,
  pv = NULL,
  pxz = 0,
  const = "none"
)
}
\arguments{
\item{N}{sample size.}

\item{changepoints}{borders of time intervals.}

\item{baseline}{baseline rate in intervals.}

\item{beta}{effect of X (log hazard ratio).}

\item{gamma}{effect of Z (log hazard ratio).}

\item{delta}{effect of V (log hazard ratio).}

\item{comprisk}{rate of competing risk event.}

\item{alpha}{effect of X on competing risk (log hazard ratio).}

\item{zeta}{effect of Z on competing risk (log hazard ratio).}

\item{px}{probability of X in intervals.}

\item{pz}{probability of Z in intervals.}

\item{pv}{probability vector for categories of V.}

\item{pxz}{effect of Z on X (log odds ratio).}

\item{const}{Whether the distributions of X and Z 
vary over time is controlled with argument `const'.
const="XZ" means that both X and Z do not change over time.
const="X" means that only X does not change over time.
const="Z" means that only Z does not change over time.}
}
\description{
Simulate from piecewise constant exponential regression model
with two binary variables X, Z whose distributions and effects
may vary in time intervals.
}
\author{
Thomas A. Gerds <tag@biostat.ku.dk>
}
