% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simPrescriptionData.R
\name{simPrescriptionData}
\alias{simPrescriptionData}
\title{Prescription data simulation function}
\usage{
simPrescriptionData(n, m = 10, packages = list(A07 = list(c(200, 30), c(400,
  100), c(400, 300), c(500, 60)), A12B = list(c(750, 100), c(750, 250), c(75,
  500))), max.packages = 3, startDate = "1995-01-01")
}
\arguments{
\item{n}{Number of patients}

\item{m}{Maximal number of prescription dates per patient}

\item{packages}{Named list of association lists. The names of the list are ATC codes.}

\item{max.packages}{Integer. How many packages do we want as maximal on one day.}
}
\description{
Simulate prescription data
}
\details{
Simulate prescription data alike the Danish medical registry


https://www.whocc.no/atc_ddd_index/
http://www.medicinpriser.dk/
http://pro.medicin.dk/

A07: strength 200, number of pills 30
     strength 400, number of pills 100
     etc

A12: strength 750, number of pills 100
     strength 750, number of pills 250
     etc

Association list to be used as package argument 

\code{list("A07"=list(c(200,30),c(400,100),c(400,300),c(500,60)),
      "A12B"=list(c(750,100),c(750,250),c(75,500)))}
}
\examples{

set.seed(1)
simPrescriptionData(1)
set.seed(2)
simPrescriptionData(2)

}
\author{
Helene Charlotte Rytgaard \email{hely@sund.ku.dk}, Thomas Alexander Gerds \email{tag@biostat.ku.dk}
}
