% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simulate_trial.R
\name{simulate_trial}
\alias{simulate_trial}
\title{simulate_trial}
\usage{
simulate_trial(
  n,
  followup = 5,
  prop_Y = 0.5,
  prop_censor = 0.3,
  prop_compete = 0.2,
  prop_discontinue = 0.3,
  prop_switch_if_discontinue = 0.5,
  prop_T = 0.5,
  seed = 1234
)
}
\arguments{
\item{n}{Number of individuals.}

\item{followup}{Years of follow-up.}

\item{prop_Y}{Probability of outcome.}

\item{prop_censor}{Probability of censoring.}

\item{prop_compete}{Probability of competing risk.}

\item{prop_discontinue}{Probability of discontinuation.}

\item{prop_switch_if_discontinue}{Probability of switching treatment after discontinuation.}

\item{prop_T}{Probability of the random time-dependent variable `T`.}

\item{seed}{Optional random seed.}
}
\value{
A list of three datasets corresponding to baseline, time-dependent variables
(`timedep`), and outcomes.
}
\description{
Simulate data for a trial with baseline covariates, time-dependent treatment
variables, and outcomes.
}
\examples{
library(data.table)
simulate_trial(seed = 2, n = 10)

}
\author{
Christian Torp-Pedersen
}
