% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heaven-data.R
\docType{data}
\name{sks_codes}
\alias{sks_codes}
\title{SKS codes}
\format{
A data table with 27864 rows and 2 variables
\describe{
   \item{kode}{A character vector of the code}
   \item{tekst}{A description of the code (in Danish)}
   }
}
\source{
http://medinfo.dk/sks/
}
\usage{
sks_codes
}
\description{
Several sections of the Sygehusvaesnets Klassifikationssystem (SKS).
Retrieved from SKS in October 2018. Consists of B, D, K, M, and U codes.
}
\keyword{datasets}
