% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize.prodlim.R
\name{standardize.prodlim}
\alias{standardize.prodlim}
\title{Standardize absolute risks}
\usage{
standardize.prodlim(object, var1, var2, times, data, weights, ...)
}
\arguments{
\item{object}{Object obtained with \code{prodlim} stratified with
two categorical covariates.}

\item{var1}{Name of one of the two categorical variables. The
absolute risks are standardized within the levels of this
variables to the distribution of the other categorical
variable in the full data.}

\item{var2}{Name of the other of the two categorical variables,
i.e., the one on which the weights are based.}

\item{times}{Compute absolute risks at these time points}

\item{data}{Only if weights are not given. Data set in which to compute the weights for
standardization as percentage in each of the levels of \code{var2}.}

\item{weights}{A numeric vector of weights that sum to 1 and has length equal
to the number of strata defined by \code{var2}. If omitted, weights are calculated using \code{data}.}

\item{...}{not used}
}
\description{
Compute standardized absolute risks based on a stratified Kaplan-Meier or Aalen-Johansen estimate
}
\details{
Fit prodlim with two categorical covariates, say age and sex. Then
standardize the absolute risk within each of the two gender to the
age distribution in whole data set.
}
\examples{
library(survival)
library(prodlim)
library(data.table)
data(pbc)
d <- na.omit(pbc[,c("time","status","trt","age","sex")])
d$agegroup <- cut(d$age,breaks=c(0,40,50,60,100))
fit <- prodlim(Hist(time,status==1)~trt+agegroup,data=d)
standardize.prodlim(fit,var1="trt",var2="agegroup",data=d,times=1000)
standardize.prodlim(fit,var1="trt",var2="agegroup",data=d,times=c(1000,1500))
}
\author{
Thomas A. Gerds <tag@biostat.ku.dk>
}
