% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize.proportion.R
\name{standardize.proportion}
\alias{standardize.proportion}
\title{Standardize proportions and absolute risks to a given age distribution}
\usage{
standardize.proportion(vars, agevar, byvar, data, ref = 2, formula, cause,
  times, level = 0.95, ...)
}
\arguments{
\item{vars}{Name vector of categorical variables}

\item{agevar}{Name of categorical age variable}

\item{byvar}{Name of categorical strata variable}

\item{data}{Data set}

\item{ref}{For categorical variable that are not factors the level for which we calculate the standardized proportion}

\item{formula}{Passed to prodlim for calculation of absolute risks}

\item{cause}{In case of competing risks the cause of interest}

\item{times}{The time point(s) for evaluation of absolute risks}

\item{level}{Confidence level}

\item{...}{Not (yet) used}
}
\value{
Data table with standardized proportions and standardized absolute risks
}
\description{
Standardize proportions and absolute risks to a given age distribution
}
\details{
Standardize proportions and absolute risks to a given age distribution
}
\examples{
library(riskRegression)
set.seed(84)
d <- riskRegression::sampleData(83)
d[,X1:=factor(X1,levels=c("0","1"),labels=c("0","1"))]
d[,X2:=factor(X2,levels=c("0","1"),labels=c("0","1"))]
d[,agegroups:=cut(X6,c(-Inf,40,60,Inf))]
standardize.proportion(vars=c("X1","X2"),agevar="agegroups",byvar="X4",data=d)

}
\seealso{
standardize.prodlim
}
\author{
Thomas A. Gerds <tag@biostat.ku.dk>
}
