% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/subtract_procedures.R
\name{subtract_procedures}
\alias{subtract_procedures}
\title{Extraction of simple procedures}
\usage{
subtract_procedures(data, time.frame = NULL, procedure = NULL,
  incombination = NULL, record.id = "RECNUM", procedure.codes = "OPR",
  out.date = "ODTO")
}
\arguments{
\item{data}{Input data with containing procedures}

\item{time.frame}{An integer value setting time-window in number of days between procedures in 
order to denote them as simple. Default is one day.}

\item{procedure}{A list of procedure codes to be extracted. 
Default is c('KFNG','KFN','KFK','KFM','KFG','KFJ', 'KFP','UXU', 'UXCC00A','UXCC00')
corresponding to PCI, CABG, klapoperation, DC konvertering, arbejds EKG, 
CT hjerte and CT Thorax}

\item{incombination}{indikator of whether combined or simple procedures are sought. Default is Null.}

\item{record.id}{Name of patient identifier. Default is RECNUM corresponding to the name in OPR.}

\item{procedure.codes}{Name of procedure codes}

\item{out.date}{Name of procedure date. Default is ODTO corresponding to the name in OPR.}
}
\value{
A list containg the filtered data.
}
\description{
Extracting simple procedures given from a list of procedure codes from a data file.
             The procedures must be the only procedure occuring within a given time-frame.
}
\examples{
\dontrun{
Generating test-data that are OPR a like.
RECNUM = 1:20
atc = rep(c("KFN9884","ZZl0986","UX9004","UXCC001", "UXCC00A", "KFNG978"), each =50)
eksd = rep(c("2002-01-15","2002-01-16","2004-02-15","2002-02-17", "2004-02-18"), each =10)

testdat <- data.table(RECNUM, atc, eksd) 


} 
}
