% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/trialSimulation.R
\name{trialSimulation}
\alias{trialSimulation}
\title{trialSimulation}
\usage{
trialSimulation(seed=123,n=1000)
}
\arguments{
\item{seed}{is random seed}

\item{n}{is number of individuals}
}
\value{
The function returns a list of three datasets corresponding to baseline,
time dependent variables (dep_vars) and outcomes
}
\description{
This function was generated by CHATGPT based on the following request:
I need a program to produce data for a simulated trial of n individuals. 
It should have baseline data of sex (equal distribution) and age (between 50 
and 80 years).  At baseline individuals are randomly given treatment A or B.  
During follow-up of five years 50% have the outcome Y (0/1), 30% are censored 
at varying times and the final 20% have the competing risk of death. All 
three outcomes occur randomly during the trial except for treatment A that 
reduces the risk of outcome by 50% when the actual treatment is A.  
During follow-up at random times 30% discontinue the allocated treatment 
and half of these switch to ther other treatment.  There is further another 
time dependent variable T which has no effect on outcomes, but which occur 
for 50% of indivuals during about a year sometime during the five year period.

The product should be three datasets.  The baseline  data have age, sex 
and baseline treatment.  The outcome dataset should have four variables: 
ID, Y, Censor and Compete - For the columns of Censor, 
Compete and Y the value is the time of the outcome.  The final dataset is 
the time dependent variables of the two treatments and the extra variable T.  
This dataset should have ID, start, end, value=1 and variable (T, A, B)
}
\examples{
library(data.table)
trialSimulation(seed=2,n=10)
}
\author{
Christian Torp-Pedersen
}
