% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{u_codes}
\alias{u_codes}
\title{Danish diagnostic procedure codes}
\format{A data table with 590 rows and 2 variables
\describe{
   \item{kode}{A character vector of the code}
   \item{tekst}{A description of the procedure (in Danish)}
   }}
\source{
http://medinfo.dk/sks/
}
\usage{
u_codes
}
\description{
The classification system for diagnostic procedures as used in SKS. Retrieved from SKS in October 2018
}
\keyword{datasets}
