% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/use_drake_project.R
\name{use_drake_project}
\alias{use_drake_project}
\title{Sets up your directory for a drake project}
\usage{
use_drake_project()
}
\value{
Modifies your directiory.
}
\description{
Helper function that will create some scafolding to help you
 easily start a drake project
}
\author{
Matthew Phelps
}
