% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/whatToDo.R
\name{whatToDo}
\alias{whatToDo}
\title{What to do: split and combine or bootstrap?}
\usage{
whatToDo(FUN, B = 10, K = 10, mydata, myargs, myargsBoot, myargsPart,
  timeunit = "mins", myseed = 29021986)
}
\arguments{
\item{FUN}{Name of function to call.}

\item{B}{Number of bootstrap replications.}

\item{K}{Number of splits.}

\item{mydata}{Data.}

\item{myargs}{Common arguments of the function for both Bootstrap and split method.}

\item{myargsBoot}{Specific arguments of the function for the Bootstrap method.}

\item{myargsPart}{Specific arguments of the function for the Bootstrap method.}

\item{timeunit}{Time units (to print the results).}

\item{myseed}{Random seed (for reproducibility).}
}
\value{
nothing
}
\author{
Paul Blanche
}
