% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/charlsonIndex.R
\name{charlsonIndex}
\alias{charlsonIndex}
\title{charlsonindex - define charlson index at a selected time}
\usage{
charlsonIndex(data,ptid='pnr',vars,data.date,charlson.date,look.back=5
,ccodes=charlson.codes)
}
\arguments{
\item{data}{- A dataframe/table with disease codes and dates}

\item{ptid}{- Variable defining individual}

\item{vars}{- variables where disease codes should be searched for}

\item{data.date}{- variable in data with date of disease}

\item{charlson.date}{- variable defining index date for determination of 
charlson index}

\item{look.back}{- numnber of years to look back from charlson.date for 
diseases}

\item{ccodes}{- named list of charlson codes. Default uses list 
supplied by heaven}
}
\value{
A list with 2 data.tables - index holding charlson index and elements 
holding presence of each category
}
\description{
This function uses 19 disease categories and their weights to establish
charlson index at a particular time indicated by a variable

The suggested list of Charlson definitions comes partly from 
Quann (Am J Epidemiol.2011 173:676-82), partly from DMCG.dk benchmarking 
consortium and with further minor adjustments provided by Peter Ascanius 
Jacobsen.  It is unlikely that any international list of definitions will 
apply for a Danish Population and the listing used should be either cited or
provided in a supplement for publication.
}
\details{
The charlson weights and selection of disease codes are from DCMG.dk
}
\examples{
require(data.table)
set.seed(211)
adm <- simAdmissionData(10000)
adm[,charlson.date:=as.Date("2017-01-01")]
ci <- charlsonIndex(adm,ptid='pnr',vars='diag',data.date='inddto',
  charlson.date="charlson.date")
}
\author{
Christian Torp-Pedersen
}
