% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/checkMicroData.R
\name{checkMicroData}
\alias{checkMicroData}
\title{Checks univariate tables for microdata}
\usage{
checkMicroData(dat, min.val = 5)
}
\arguments{
\item{dat}{Table to be checked. Must be the summarized output of Publish::utable()}

\item{min.val}{The minimum value allowed, anything below this value will be flagged as microdata}
}
\value{
A message showing the location of any microdata found, or a statement that no microdata was found
}
\description{
Checks univariate tables for microdata
}
\examples{
checkMicroData(summary(Publish::utable(cyl ~ vs, data=mtcars)))
}
