% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{d_codes}
\alias{d_codes}
\title{Danish ICD-10 codes}
\format{A data table with 20739 rows and 2 variables
\describe{
   \item{kode}{A character vector of the code}
   \item{tekst}{A description of the diagnosis (in Danish)}
   }}
\source{
http://medinfo.dk/sks/
}
\usage{
d_codes
}
\description{
The Danish version of the ICD-10 codes, part of the Sygehusvaesnets Klassifikationssystem (SKS). Retrieved from SKS in October 2018
}
\keyword{datasets}
