% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dpp.R
\name{dpp}
\alias{dpp}
\title{Create data preprocessing object (dpp)}
\usage{
dpp()
}
\description{
Create data preprocessing object
}
\details{
Create data preprocessing object (dpp)
}
\author{
Helene Charlotte Rytgaard
}

