% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractCode.R
\name{extractCode}
\alias{extractCode}
\title{Extraction of diseases by diagnoses}
\usage{
extractCode(dat, disease = NULL, inclusions = NULL, exclusions = NULL,
  p.in = NULL, p.out = NULL, mult = FALSE, keep = "", pat = NULL,
  prefix = "", entryvar = "inddto", id = "pnr", codevar = "diag",
  patvar = "pattype", record.id = "recnum", indexvar = NULL,
  index.int = NULL, outcome = FALSE, lmdb = FALSE)
}
\arguments{
\item{dat}{Dataset containing diagnoses, dates, patient id, patient type etc.}

\item{disease}{Characterstring containing pre-specified name of diseases. See \href{../doc/predefined_diseases.pdf}{definitions of diseases}.}

\item{inclusions}{Characterstring, where additional diagnoses can be included. If disease is not defined, inclusions will be the extracted diagnoses.}

\item{exclusions}{Characterstring, specifying diagnoses to be omitted.}

\item{p.in}{Date of period p.in, as characterstring in the format "YYYY-MM-DD".}

\item{p.out}{Date of period p.out, as characterstring in the format "YYYY-MM-DD".}

\item{mult}{Logical. Specifies if multiple diseases are chosen.}

\item{keep}{Takes the values "first" or "last". Specifies if only the first or last record of each patient should be output. 
If mult=TRUE and keep="first" the first diagnosis of each disease specified will be extracted.}

\item{pat}{Number or vector defining types of patients to include (pattype: 0,1,2,3), default is all types.}

\item{prefix}{Character string of prefix name for the resulting date variable of disease.}

\item{entryvar}{Name of the variable in data that contains the entrydate of diagnosis.}

\item{id}{Name of the variable in data that contains patient id.}

\item{codevar}{Name of the variable in data that contains diagnoses.}

\item{patvar}{Name of the variable in data that contains the type of patient.}

\item{record.id}{Name of the variable in data that contains the record number for each patient.}

\item{indexvar}{Name of the variable in data that contains the index date.}

\item{index.int}{Number of days before index date to search for specific diagnoses.}

\item{outcome}{Logical. If true, all specific diagnoses after index date is returned. If set to TRUE, do not specify index.int.}

\item{lmdb}{Logical. If true, data with atc codes is expected, and record.id and patvar are not expected to be in the data.}
}
\value{
A list of three elements. data: the extracted data. diagnoses: contains the diagnoses specified 
to be extracted. unique.icd: contains every unique diagnosis extracted. If exclusions are specified, these are contained in an additional list element: excl.
}
\description{
Filtering of LPR registry data according to a given set of diseases
}
\details{
Extracts specific selected or predefined diagnoses. If specified by keep only the first or last occurrence of the diagnoses are extracted.
Diagnoses in a specific period of time can also be extracted by p.in and p.out.
}
\examples{
\dontrun{

# Simulated LPR-registry data
data(lpr.data)

# Extract diagnoses related to myocardial infarction (mi) for patients of type 1, after 01-01-2007.
dat.extracted <- extractCode(lpr.data,disease=c("mi"),pat=1,p.in='2007-01-01')

# View first 6 lines of extracted data 
head(dat.extracted$data)

# Codes specified to be extracted
dat.extracted$diagnoses

# Unique codes extracted
dat.extracted$unique.icd

# Extract diagnoses related to myocardial infarction (mi) within the period 01-01-2007 and 12-31-2008.
dat.extracted <- extractCode(lpr.data,disease=c("mi"),p.in='2007-01-01',p.out='2008-12-31')

# Extract diagnoses related to bleeding (without hemorrhagic stroke) or atrial fibrilation, and only include first
# diagnosis of each disease for each patient.
dat.extracted <- extractCode(lpr.data,disease=c("bleeding","af"),mult=T,keep='first',prefix='diag')

# Extract all diagnoses p.ining with an 'A', and exclude all diagnoses with 'A21' and 'A7'.
dat.extracted <- extractCode(lpr.data,inclusions='A',exclusions=c('A21','A7'),prefix='a')

# Extract ischemic heart disease without I21 and I22:
dat.extracted <- extractCode(lpr.data,disease='ihd',exclusions=c('I21','I22'),prefix='ihd_excl_ami')

# Extract ischemic heart disease diagnoses after index date
dat.extracted <- extractCode(lpr.data,disease=c("ihd"),p.in='2007-01-01',p.out='2008-12-31',indexvar='indexdate',outcome = TRUE)

# Extract ischemic heart disease diagnoses within 30 days before index date
dat.extracted <- extractCode(lpr.data,disease=c("ihd"),p.in='2007-01-01',p.out='2008-12-31',indexvar='indexdate',index.int=30)
}
}
\author{
Regitze Kuhr Skals <r.skals@rn.dk>
}
