% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/extractCode.R
\name{extractCode}
\alias{extractCode}
\title{Extraction of diseases by diagnoses}
\usage{
extractCode(dat, disease = NULL, inclusions = NULL,
  exclusions = NULL, p.in = NULL, p.out = NULL, keep = "",
  pat = NULL, prefix = "", entryvar = "inddto", id = "PNR",
  codevar = "diag", patvar = "pattype", indexvar = NULL,
  index.int = NULL, lmdb = FALSE)
}
\arguments{
\item{dat}{Dataset containing diagnoses/atc codes, entrydate, patient id, patient type, record number and optionally index date. If extraction of atc codes is desired, patient type and record number is not expected in the input data (see lmdb).}

\item{disease}{Characterstring containing pre-specified name of diseases. See \href{../doc/predefined_diseases.pdf}{definitions of diseases}.}

\item{inclusions}{Characterstring, where additional diagnoses/atc codes can be included. If disease is not specified, inclusions will be the extracted codes.}

\item{exclusions}{Characterstring, specifying codes to be omitted. If disease and inclusions are not specified, all codes except exclusions will be extracted.}

\item{p.in}{Date of period start, if a specific period of time is desired. Characterstring in the format "YYYY-MM-DD".}

\item{p.out}{Date of period end. See p.in.}

\item{keep}{Takes the values "first" or "last". Specifies if only the first or last record of each patient should be output. 
If multiple diseases are chosen and keep="first" the first code of each disease specified will be extracted.}

\item{pat}{Number or vector defining types of patients to include (pattype: 0,1,2,3), default is all types.}

\item{prefix}{Character string of prefix name for the resulting date variable of disease. If not specified, and disease is specified, the name in disease is chosen as prefix.}

\item{entryvar}{Name of the variable in data that contains the entrydate/prescription date of a diagnosis/atc code.}

\item{id}{Name of the variable in data that contains patient id.}

\item{codevar}{Name of the variable in data that contains diagnoses or atc codes.}

\item{patvar}{Name of the variable in data that contains the type of patient.}

\item{indexvar}{Name of the variable in data that contains the index date.}

\item{index.int}{Numeric. Number of days before or after an index date to search for specific diagnoses/atc codes. If negative, days before index date. If positve, days after index date (i.e. outcome).}

\item{lmdb}{Logical. If true, data with atc codes is expected, and and patvar are not expected to be in the data.}
}
\value{
A list of three elements. data: the extracted data. codes: contains the codes specified 
to be extracted. unique.codes: contains every unique code extracted. If exclusions are specified, these are contained in an additional list element: excl.
}
\description{
Filtering of LPR registry data according to a given set of diseases, or filtering of medical prescription data according to given atc codes.
}
\details{
Extracts specific selected ICD- or ATC codes, or predefined diseases by diagnoses. If specified by keep, only the first or last occurrence of the code is extracted.
}
\examples{
\dontrun{

# Simulated LPR-registry data
set.seed(2976)
lpr.data <- simAdmissionData(100)

# Extract diagnoses related to heart failure
dat.extracted <- extractCode(lpr.data,disease=c("hf"),id='pnr')

# View first 6 lines of extracted data 
head(dat.extracted$data)

# Codes specified to be extracted
dat.extracted$codes

# Unique codes extracted
dat.extracted$unique.codes

# Extract diagnoses related to cancer after the date 01-01-2007.
dat.extracted <- extractCode(lpr.data,disease=c("cancer"),p.in='2007-01-01',id='pnr')

# Extract diagnoses related to liver disease or cancer.
dat.extracted <- extractCode(lpr.data,disease=c("liver","cancer"),prefix='diag',id='pnr')

# Extract cancer diagnoses, and keep only diagnosis of patients of type 3
dat.extracted <- extractCode(lpr.data,disease="cancer",pat=3,id='pnr')

# Extract all diagnoses begining with 'DI', and exclude all diagnoses with 'DI22' and 'DI9'.
dat.extracted <- extractCode(lpr.data,inclusions='DI',
                             exclusions=c('DI22','DI9'),prefix='i',id='pnr')

# Extract ischemic heart disease diagnoses within three years after index date
dat.extracted <- extractCode(lpr.data,inclusions=c("DI"),indexvar='indexdate',
                              index.int=365*3,id='pnr')

# Extract bleeding diagnoses within one year before index date
dat.extracted <- extractCode(lpr.data,inclusions=c("DN"),indexvar='indexdate',
                             index.int=-365,id='pnr')
}
}
\author{
Regitze Kuhr Skals <r.skals@rn.dk>
}
