% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/Framingham.R
\docType{data}
\name{Framingham}
\alias{Framingham}
\title{Documentation for data Framingham}
\format{An object of class \code{data.table} (inherits from \code{data.frame}) with 11627 rows and 39 columns.}
\usage{
Framingham
}
\description{
The Framingham dataset is a public version of several examinations of people
in the Framingham study.  Various variables have been adjusted to ensure
anonymity of people. Thus, the data cannot be used for science, only demon-
stration.
}
\details{
Details of variables can be found in the inst/doc folder: framdoc.pdf
}
\keyword{datasets}
