% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heaven-package.R
\docType{data}
\name{Framingham}
\alias{Framingham}
\title{Public version of Framingham data}
\format{The Framingham dataset is a public version of several examinations of people in the Framingham study. Various variables have been adjusted to ensure anonymity of people. Thus, the data cannot be used for science, only demonstration of R-code. 
\describe{
\item{randid}{person code} 
\item{sex}{1=male 2=female} 
\item{totchol}{cholesterol mg/dL} 
\item{age}{years} 
\item{sysbp}{mm Hg} 
\item{diabp}{mm Hg} 
\item{cursmoke}{0=no 1=yes} 
\item{cigpday}{0=no 1-90 per day} 
\item{bmi}{body mass index} 
\item{diabetes}{0=no 1=yes} 
\item{bpmeds}{0=no 1=yes} 
\item{heartrte}{beats per minute} 
\item{glucose}{mg/dL} 
\item{educ}{education} 
\item{prevchd}{Prevalent coronary disease} 
\item{prevap}{prevalent angina} 
\item{prevmi}{prevalent MI} 
\item{prevstrk}{prevalent stroke} 
\item{prevhyp}{prevalent hypertension} 
\item{time}{days since baseline} 
\item{period}{exmination cycle} 
\item{hdlc}{hdl cholesterol} 
\item{ldlc}{ldl cholesterol} 
\item{death}{0=no 1=yes} 
\item{angina}{pectoris} 
\item{hospmi}{hospitalised MI} 
\item{MI_FCHD}{hospitalised MI or fatal coronary disease} 
\item{anychd}{any coronary heart disease} 
\item{stroke}{0=no 1=yes} 
\item{cvd}{cardiovascular disease} 
\item{hyperten}{0=no 1=yes} 
\item{timeap}{days til angina} 
\item{timemi}{days til MI} 
\item{timemifc}{days til MI or fatal coronary} 
\item{timechd}{days til coronary disease} 
\item{timestrk}{days til stroke} 
\item{timecvd}{days til cardiovascular disease} 
\item{timedth}{days til death} 
\item{timehyp}{days til hypertension} 
}}
\description{
The Framingham dataset is a public version of several examinations of people in the Framingham study. Various variables have been adjusted to ensure anonymity of people. Thus, the data cannot be used for science, only demonstration of R-code.
}
\examples{

data(Framingham)


}
\keyword{datasets}
