% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAdmLimits.R
\name{getAdmLimits}
\alias{getAdmLimits}
\title{Get Admission time}
\usage{
getAdmLimits(
  dt,
  pnr = "pnr",
  inddto = "inddto",
  uddto = "uddto",
  keep = NULL,
  collapse = FALSE,
  error = "warn"
)
}
\arguments{
\item{dt}{data.table containing the variables}

\item{pnr}{Variable with ID for each subject/group (default is \code{pnr}).}

\item{inddto}{Variable with entry times. Must be date or numeric (default is \code{inddto}).}

\item{uddto}{Variable with exit times. Must be date or numeric (default is \code{uddto}).}

\item{keep}{Vector of names of variables of \code{dt} to keep in the output}

\item{collapse}{if \code{TRUE} return only the lines with non-overlapping admissions}

\item{error}{Character. One of \code{"warn"}, \code{"remove"} remove lines with errors,
\code{"flag"} add a new variable called error. An error is defined as either a missing value in
either inddto or uddto or when inddto>uddto.}
}
\value{
data.table with two new variables first.indate and last.outdate.
}
\description{
Adds first admission date and last discharge date to a sequence of admissions. The default input data.table is the current standard
by which admissions are available in Statistics Denmark.
}
\examples{
set.seed(8)
lpr <- simAdmissionData(10)
## Variables have default names
adm <- getAdmLimits(lpr)

## with errors
lpr1 <- lpr
lpr1 <- lpr1[uddto>as.Date("2010-01-01"),uddto:=NA]
adm1 <- getAdmLimits(lpr1)
adm1 <- getAdmLimits(lpr1,error="remove")
adm1 <- getAdmLimits(lpr1,error="flag")

## case where Variables have custom names
data.table::setnames(lpr, c('pnr','inddto','uddto'), c('personid', 'Entrydate','Outdate'))
adm2 <- getAdmLimits(lpr,pnr='personid',inddto='Entrydate',uddto='Outdate')

## duplicated and overlap
set.seed(8)
lpr3 <- simAdmissionData(10)
lpr3 <- lpr3[sample(1:NROW(lpr3),replace=TRUE,size=round(NROW(lpr3)*1.5))]
adm3 <- getAdmLimits(lpr3,collapse=TRUE)

}
\author{
Thomas A. Gerds <tag@biostat.ku.dk>
}
