% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getAdmLimitsV2.R
\name{getAdmLimitsV2}
\alias{getAdmLimitsV2}
\title{Get Admission time}
\usage{
getAdmLimitsV2(dt, personid, inddto, uddto)
}
\arguments{
\item{dt}{data.table containing the variables}

\item{personid}{Variable with ID for each subject/group}

\item{inddto}{Variable with entry times. Must be date or numeric}

\item{uddto}{Variable with exit times. Must be date or numeric}
}
\value{
List containing dataset with admission dates add to data and dataset containing row errors
}
\description{
Removes row errors and returns data with add admission time
}
\examples{
\dontrun{
newpop <- getAdmLimits3(mydata,'Pnr','Entrydate','Outdate')

## List is returned containing two data.tables
newpopData <- newpop[[1]]
newpopError <- newpop[[2]]
}
}
\author{
Thomas A. Gerds <tag@biostat.ku.dk>, Peter Enemark Lund <peter.l@rn.dk>
}
