% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/halal.R
\name{halal}
\alias{halal}
\title{Extraction of diseases by diagnoses}
\usage{
halal(dat, disease = NULL, inclusions = NULL, exclusions = NULL,
  p.in = NULL, p.out = NULL, mult = FALSE, first.pnr = FALSE,
  pat = NULL, prefix = "", entryvar = "inddto", id = "pnr",
  outvar = "uddto", codevar = "diag", patvar = "pattype",
  record.id = "recnum")
}
\arguments{
\item{dat}{Data with diagnoses.}

\item{disease}{Characterstring containing pre-specified name of disease after HALAL definitions. Can be one or more of: "af",
"rheum_heart_valve_disease", "dvt_pe","stroke_tci_systemic_embolism_thrombosis",
"bleeding_without_hemstroke","cardiomyopathy", "hf", "lung_edema", "HF", "ihd","ap","ami","perart",
"cancer","chronic_kidney_disease","copd","liver_disease","alcohol". See \href{../doc/halal_defs.pdf}{definitions of diseases}.}

\item{inclusions}{Characterstring, where additional diagnoses can be included. If disease is not defined, inclusions will be the extracted diagnoses.}

\item{exclusions}{Characterstring, specifying diagnoses to be omitted.}

\item{p.in}{Date of period start.}

\item{p.out}{Date of period end.}

\item{mult}{Logical. Specifies if multiple diseases have been chosen.}

\item{first.pnr}{Logical. Specifies if only the first record of each patient should be output. 
If mult=TRUE and first.pnr=TRUE the first diagnosis of each disease specified will be extracted.}

\item{pat}{Number or vector defining types of patients to include (pattype: 0,1,2,3), default is all types.}

\item{prefix}{Character string of prefix name for the resulting date variable of disease.}

\item{entryvar}{Name of the variable in data that contains the entrydate of diagnosis.}

\item{id}{Name of the variable in data that contains patient id.}

\item{outvar}{Name of the variable in data that contains the exitdate of diagnosis.}

\item{codevar}{Name of the variable in data that contains diagnoses.}

\item{patvar}{Name of the variable in data that contains the type of patient.}

\item{record.id}{Name of the variable in data that contains the record number for each patient.}
}
\value{
A list of three elements. The first one is the extracted data. The second element contains the diagnoses specified 
to be extracted. The third element contains every unique diagnosis extracted.
}
\description{
Filtering of LPR registry data according to a given set of diseases
}
\details{
Extracts specific selected or predefined diagnoses. If specified by first.pnr only the first occurrence of the diagnoses are extracted.
Diagnoses in a specific period of time can also be extracted by p.in and p.out.
}
\examples{
\dontrun{

# Creating simulated LPR-registry data
alphabet <- LETTERS[seq( from = 1, to = 26 )]
abc <- sample(alphabet,500000,replace=TRUE)
numbers <- c(100:999)
icd_codes <- sample(numbers, 500000, replace = T)
diag <- paste(abc,icd_codes,sep='')
diag <- paste('D',diag,sep='')
inddto <- c(11688:14974)
inddto <- as.Date(sample(inddto,500000,replace=TRUE),origin='1970-01-01')
uddto <- c(1:10)
uddto <- inddto+sample(uddto,500000,replace = T)
pattype <- sample(c(1:3),500000,replace = T)
pnr <- sample(c(1:300000),500000,replace=TRUE)
pnr <- sort(pnr)
lpr.data <- data.frame(pnr,diag,inddto,uddto,pattype,recnum=rep(1,500000))

# Extract diagnoses related to acute myocardial infarction (ami) for patients of type 1, after 01-01-2007.
dat.extracted <- halal(lpr.data,disease=c("ami"),pat=1,p.in=as.Date('01012007',format='\%d\%m\%Y'),prefix='ami')

# View first 6 lines of extracted data 
head(dat.extracted$data)

# Codes specified to be extracted
dat.extracted$diagnoses

# Unique codes extracted
dat.extracted$unique.icd

# Extract diagnoses related to acute myocardial infarction (ami) within the period 01-01-2007 and 12-31-2008.
dat.extracted <- halal(lpr.data,disease=c("ami"),p.in=as.Date('01012007',format='\%d\%m\%Y'),p.out=as.Date('31122008',format='\%d\%m\%Y'),prefix='ami')

# Extract diagnoses related to bleeding (without hemorrhagic stroke) or atrial fibrilation, and only include first
# diagnosis of each disease for each patient.
dat.extracted <- halal(lpr.data,disease=c("bleeding_without_hemstroke","af"),mult=TRUE,first.pnr=TRUE,prefix='diag')

# Extract all diagnoses starting with an 'A', and exclude all diagnoses with 'A21' and 'A7'.
dat.extracted <- halal(lpr.data,inclusions='A',exclusions=c('A21','A7'),prefix='a')

# Extract ischemic heart disease without I21 and I22:
dat.extracted <- halal(lpr.data,disease='ihd',exclusions=c('I21','I22'),prefix='ihd_excl_ami')
}
}
\author{
Regitze Kuhr Skals <r.skals@rn.dk>
}
