% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/halal.R
\name{halal}
\alias{halal}
\title{Extraction of diseases by diagnoses}
\usage{
halal(dat, disease = NULL, inclusions = NULL, exclusions = NULL,
  p.in = NULL, p.out = NULL, first.pnr = FALSE, pat = NULL,
  prefix = "", entryvar = "inddto", id = "pnr", outvar = "uddto",
  codevar = "diag", patvar = "pattype", record.id = "recnum")
}
\arguments{
\item{dat}{Data with diagnoses.}

\item{disease}{Characterstring containing pre-specified name of disease after HALAL definitions. Can be one or more of: "af",
"rheum_heart_valve_disease", "dvt_pe","stroke_tci_systemic_embolism_thrombosis",
"bleeding_without_hemstroke","cardiomyopathy", "hf", "lung_edema", "HF", "ihd","ap","ami","perart",
"cancer","chronic_kidney_disease","copd","liver_disease","alcohol".}

\item{inclusions}{Characterstring, where additional diagnoses can be included. If disease is not defined, inclusions will be the extracted diagnoses.}

\item{exclusions}{Characterstring, specifying diagnoses to be omitted.}

\item{p.in}{Date of period start.}

\item{p.out}{Date of period end.}

\item{first.pnr}{Logical. Specifies if only the first record of each patient should be output.}

\item{pat}{Number or vector defining types of patients to include (pattype: 0,1,2,3), default is all types.}

\item{prefix}{Character string of prefix name for the resulting date variable of disease.}

\item{entryvar}{Name of the variable in data that contains the entrydate of diagnosis.}

\item{id}{Name of the variable in data that contains patient id.}

\item{outvar}{Name of the variable in data that contains the exitdate of diagnosis.}

\item{codevar}{Name of the variable in data that contains diagnoses.}

\item{patvar}{Name of the variable in data that contains the type of patient.}

\item{record.id}{Name of the variable in data that contains the record number for each patient.}
}
\author{
Regitze Kuhr Skals
}
