% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/heaven-package.R
\docType{package}
\name{heaven-package}
\alias{heaven}
\alias{heaven-package}
\title{heaven: Tools for Statitics Denmark}
\description{
Fast and furious implementations of often used data preparation steps. These functions are adapted to Danish registry data. Designed to be used with Statistics Denmark
}
\keyword{internal}
