% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypertensionBaseline.R
\name{hypertensionBaseline}
\alias{hypertensionBaseline}
\title{Hypertension at baseline}
\usage{
hypertensionBaseline(data, pnr = "pnr", atc = "atc", eksd = "eksd",
<<<<<<< HEAD
  date = "date")
=======
  baseline = "baseline")
>>>>>>> 6b35cf84506e4e5490ad8028cc0d1c987cfee4be
}
\arguments{
\item{data}{Data set with drugs indicated by atc codes and date of drugs, e.g. lmdb}

\item{pnr}{Variable with ID for each subject/group. Default name: pnr}

\item{atc}{Variable with atc codes. Must be type character. Default name: atc}

\item{eksd}{Variable with dates. Must be type Date or numeric. Default name: eksd}

<<<<<<< HEAD
\item{date}{Variable with dates. Must be type Date or numeric. Default name: date}
=======
\item{baseline}{Variable with dates. Must be type Date or numeric. Default name: baseline}
>>>>>>> 6b35cf84506e4e5490ad8028cc0d1c987cfee4be
}
\value{
List containing two datasets. First contains data with hypertension information. 
The second contains all rows with missing data that were removed before hypertension where calculated.
The variable hypertension at baseline is one if the person has received two or more types of anti-hypertensive
medications within 180 days before the baseline date (and 7 days after due to data errors). 
The codes for hypertension are currently defined in the function.
}
\examples{
<<<<<<< HEAD
Warning: examples are outdated. New ones are coming
=======
#Required library
>>>>>>> 6b35cf84506e4e5490ad8028cc0d1c987cfee4be
library(data.table)

#Load data
data(Hypertension.data) 

ht<-hypertensionBaseline(Hypertension,pnr='myID',atc='myAtc',eksd='myEksd',baseline='myBaseline')

List is returned containing two data.tables
htData  <- ht[[1]]
htError <- ht[[2]]

}
\author{
Helle Hoejmark Eriksen <helle.e@rn.dk>
}
