% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hypertensionBaseline.R
\name{hypertensionBaseline}
\alias{hypertensionBaseline}
\title{Hypertension at baseline}
\usage{
hypertensionBaseline(data, pnr = "pnr", atc = "atc", eksd = "eksd",
  baseline = "baseline")
}
\arguments{
\item{data}{Data set with drugs indicated by atc codes and date of drugs, e.g. lmdb}

\item{pnr}{Variable with ID for each subject/group. Default name: pnr}

\item{atc}{Variable with atc codes. Must be type character. Default name: atc}

\item{eksd}{Variable with dates. Must be type Date or numeric. Default name: eksd}

\item{baseline}{Variable with dates. Must be type Date or numeric. Default name: baseline}
}
\value{
List containing two datasets. First contains data with hypertension information. 
The second contains all rows with missing data that were removed before hypertension where calculated.
The variable hypertension at baseline is one if the person has received two or more types of anti-hypertensive
medications within 180 days before the baseline date (and 7 days after due to data errors). 
The codes for hypertension are currently defined in the function.
}
\examples{
#Required library
library(data.table)

#Load data
\dontrun{
data(Hypertension) 
ht<-hypertensionBaseline(data=Hypertension,pnr='myID',atc='myAtc',eksd='myEksd',baseline='myBaseline')

# A list is returned which contains two data.tables
htData  <- ht[[1]]
htError <- ht[[2]]
}
}
\author{
Helle Hoejmark Eriksen <helle.e@rn.dk>
}
