% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/kommuneRegion.R
\name{kommuneRegion}
\alias{kommuneRegion}
\title{kommuneRegion - Converts kommune number to Region}
\usage{
kommuneRegion(x)
}
\arguments{
\item{x}{3-digit kommune number}
}
\description{
Converts 3-digit kommune number to text representing one of the five 
administrative regions.
}
\details{
This function converts the 3-digit kommune number to a text corresponding to
which of the five regions in Denmark a kommune belongs to. The function is
adapted from a DST SAS-format and the conversions are valid from 2007. There 
is no very easy way to convert older kommune-numbers to the current regions
since some kommune were split in the processof developing the five regions
}
\examples{
kommuneRegion(410) # Sydjylland
}
\author{
Christian Torp-Pedersen
}
