% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lexisFromTo.R
\name{lexisFromTo}
\alias{lexisFromTo}
\title{lexisFromTo}
\usage{
lexisFromTo(indat,splitdat,invars,splitvars)
}
\arguments{
\item{indat}{- base data with id, start, end, event and other data - possibly 
already split}

\item{splitdat}{- Data with splitting guide - id/from/to/value/name}

\item{invars}{- vector of colum names for id/entry/exit/event - in that 
order, 
example: c("id","start","end","event")}

\item{splitvars}{- vector of column names containing dates to split by.
example: c("id","start","end","value","name") - must be in that order!}
}
\value{
The function returns a new data table where records have been split according 
to the splittingguide dataset. Variables
unrelated to the splitting are left unchanged.
}
\description{
lexixFromTo is a specialised version of lexis splitting.  In handling drug treatment, employment
and similar factors the individual may move in and out of conditions.  This function splits 
according such a situation.  The input is first "base data" id/in/out/event (and other variables) that
may already have been split by other functions.  The other data is a sequence of 
id/from/to/value/name - that may represent multiple conditions with from/to and with "name" to
distinguish.
}
\details{
The input to this function are two data.tables and two lists of the critical 
variables.  The BASE data it the data to be split.
This data must have a variable to identify participants, start/end times and 
a variable to indicate event after last interval.
The other table (SPLITTINGUIDE) contains possibly multiple records for each 
participants with id/from/to/value/name.
The program checks that intervals are not negative and that intervals for 
each "name" and each individual do not overlap.  Violation stops the program 
with error. Overlaps may occur in real situations, but the user needs to make 
decisions regarding this prior to this function.

It is required that the splittingguide contains at least one record.  
Missing data for key variables are not allowed and will cause errors.
}
\examples{
library(data.table)
dat <- data.table(id=c("A","A","B","B"),
                 start=c(0,100,0,100),
                 end=c(100,200,100,200),
                 event=c(0,1,0,0))
split <- data.table (id=c("A","A","A","A","B","B","B"),
                    start=c(0,50,25,150,110,150,400),
                    end= c(25,75,150,151,120,250,500),
                    value=c(1,2,3,4,1,2,3),
                    name=c("d1","d1","d2","d2","d1","d1","d2"))
#Show the dataset:
dat[]
split[]                   
temp <- lexisFromTo(dat # inddato with id/in/out/event
                   ,split # Data with id and dates
                   ,c("id","start","end","event") #names of id/in/out/event - in that order
                  ,c("id","start","end","value","name")) #Nmes var date-vars to split by
temp[]
}
\author{
Christian Torp-Pedersen
}
