% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/lexisTwo.R
\name{lexisTwo}
\alias{lexisTwo}
\title{lexisTwo}
\usage{
lexisTwo(indat,splitdat,invars,splitvars)
}
\arguments{
\item{indat}{- base data with id, start, end, event and other data - possibly already split}

\item{splitdat}{- Data with splitting guide - id and columns with dates to split by}

\item{invars}{- vector of colum names for id/entry/exit/event - in that order, 
example: c("id","start","end","event")}

\item{splitvars}{- vector of column names of columns containing dates to split by.
example: c("date1","date2","date3","date4")
The name of the id column must be the same in both datasets}
}
\value{
The function returns a new data table where records have been split according to the splittingguide dataset. Variables
unrelated to the splitting are left unchanged.
}
\description{
lexixTwo is a specialised version of lexis splitting.  While preparing data for time dependent analyses with e.g.
Cox org Poisson regression records need to be "split" at selected times.  For practical use there are often
multiple conditions such as comorbidities that will cause either a single split or no splitting.  This function
is designed to receive base data with start/end/event and another set of data where patient id is accompanied with 
one column per condition with these columns containing the time where split should occur.
}
\details{
The input to this function are two data.tables and two lists of the critical variables.  The base data it the data to be split.
This data must have a variable to identify participants, start/end times and a variable to indicate event after last interval.
The other table contains a single record for each participant and columns to identify dates to split by.  After splitting all
intervals preceding the date will have a variable identified by each column with the value "0". After the date the value i "1".
In the example the columns are dat1-date4 - but it is most useful to provide names that identify the condition which changes by
the date rather than a names which indicates a date.
}
\examples{
library(data.table)
dat <- data.table(id=c("A","A","B","B","C","C","D","D"),
                start=c(0,100,0,100,0,100,0,100),
                end=c(100,200,100,200,100,200,100,200),
                event=c(0,1,0,0,0,1,0,1))
split <- data.table (id=c("A","B","C","D"),date1=c(0,50,150,300),date2=c(25,75,175,325),date3=c(30,30,30,30),
                    date4=c(0,1,0,1))
temp <- lexisTwo(dat # inddato with id/in/out/event
       ,split # Data with id and dates
       ,c("id","start","end","event") #names of id/in/out/event - in that order
       ,c("date1","date2","date3","date4")) #Names var date-vars to split by
temp[]
}
\seealso{
lexisSeq lexisFromTo
}
\author{
Christian Torp-Pedersen
}
