% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/matchReport.R
\name{matchReport}
\alias{matchReport}
\title{matchReport - report of matching from riskSetMatch}
\usage{
matchReport(dat, id, case, caseid)
}
\arguments{
\item{dat}{- data.table with result from riskSetMatch}

\item{id}{- variable with participant id}

\item{case}{- 0=control, 1=case}

\item{caseid}{- variable defining the groups of matching cases/controls}
}
\value{
Three small tables - Number of controls for cases, use/reuse of controls, use/reuse of cases
}
\description{
The function provides very simple tables of the success of finding controls and
the reuse of cases and controls
}
\details{
This function can be helpful to define matching options.  If there is excessive reuse of controls or many
cases do not find controls it may be desirable to do further rounding of matching variables.
}
\examples{
require(data.table)
event <- c(rep(0,50),rep(1,5)) 
ptid <- paste0("P",1:55)
sex <- c(rep("fem",25),rep("mal",25),"fem","fem",rep("mal",3))
age <- c(rep(c(70,80),25),70,80,70,70,80)
caseIndex <- c(seq(1,49,1),seq(5,45,10))
controlIndex <- caseIndex
library(data.table)
dat <- data.table(ptid,event,sex,age,caseIndex,controlIndex)
out2 <- RiskSetMatch("ptid","event",c("age","sex"),dat,2,caseIndex="caseIndex",
  controlIndex="controlIndex")
matchReport(out2,"ptid","event","caseid")   
}
\seealso{
riskSetMatch
}
\author{
Christian Torp-Pedersen
}
