% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/medicinMacro.R
\name{medicinMacro}
\alias{medicinMacro}
\title{Medicin macro to estimate prescription lengths and averages}
\usage{
medicinMacro(drugs, treatments = NULL, drugdb, admdb,
  drugdb.datevar = "eksd", admdb.datevars = c("inddto", "uddto"),
  id = "pnr", strength.var = "strnum", packsize.var = "packsize",
  apk.var = "apk", collapse = TRUE, splitting = FALSE)
}
\arguments{
\item{drugs}{List of drugs (see examples).}

\item{treatments}{If specified, calcations will only be performed
for these treatments.}

\item{drugdb}{data.table with (subset of) medical drugs registry}

\item{admdb}{data.table with (subset of) hospital admission registry}

\item{drugdb.datevar}{name of the date variable in \code{drugdb}. Default is \code{"eksd"}.}

\item{admdb.datevars}{names of the date variables in \code{admdb}. Default is \code{c("inddto","uddto")}.}

\item{id}{Name of variable to identify individuals. Default is
\code{"pnr"}.}

\item{strength.var}{Name of variable to identify strength. Default
is \code{"strnum"}.}

\item{packsize.var}{Name of variable to identify size of
packages. Default is \code{"packsize"}.}

\item{apk.var}{Name of variable to identify number of
packages. Default is \code{"packsize"}.}

\item{collapse}{Logical. If \code{TRUE}, collapse contiguous
exposure periods with same computed exposure strength to one
and add periods with zero exposure.}

\item{splitting}{Split the data into 10 chunks to estimate
remaining time (and possibly speed up computation).}
}
\description{
Processing function, to perform calculations according to the data and variables specified in preprocessing object.
}
\examples{
set.seed(05021992)
N=20
lmdb=simPrescriptionData(N)
lpr=simAdmissionData(N)
drug1 = list(atc=c("R03AK11","R03AL03","R03AC02","R03AC04","R03AC19",
                    "R03AL02","R03AA01","R03AC18","R03AL01"),
             maxdepot=4000,
             period=as.Date(c("1997-01-01", "2012-12-31")),
             prescriptionwindow=2,
             doses=list(value=c(750,500,400,200,75),
                        min = c(250,200,200,100,25),
                        max = c(1000,600,800,600,100),
                        def = c(750,500,400,200,75)))
drug2=list(atc=c("C01CA01","C01AA05"),
           maxdepot=4000,
           period=as.Date(c("1997-01-01", "2012-12-31")),
           prescriptionwindow=2,
           doses=list(value=c(200, 400, 500,750),
                      min = c(100, 200, 250,750),
                      max = c(300, 800, 1000,750),
                      def = c(200, 400, 500,750)))
a <- medicinMacro(drugs=list("drug1"=drug1,"drug2"=drug2),drugdb=lmdb,admdb=lpr)
}
\author{
Helene Charlotte Rytgaard, Anders Munch and Thomas Alexander Gerds based
    on the famous SAS macro by Christian Tobias Torp-Pedersen
}
