% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mm1.R
\name{mm1}
\alias{mm1}
\title{Calculates whether patients are exposed or not at a given time}
\usage{
mm1(
  drugs,
  drugdb,
  admdb,
  time.points,
  window,
  method = "default.doses",
  stash = 0,
  cap.values = TRUE,
  verbose = FALSE
)
}
\arguments{
\item{drugs}{Relevant drugs}

\item{drugdb}{Drug database}

\item{admdb}{Admission database}

\item{time.points}{Time points}

\item{window}{Time window}

\item{method}{Default doses by default}

\item{stash}{Stash}

\item{cap.values}{Overwrite a last period with previous}
}
\value{
Exposed or not at a given time
}
\description{
Part of the \code{medicinMacro}
}
\author{
Anders Munch
NB: _potential_ purchases, becuase a ptt's might not purchase anything in the time frame -- this gives a total.drug and drug.supply.days of 0
}
