% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mm1.R
\name{mm1}
\alias{mm1}
\title{Calculates whether patients are exposed or not at a given time}
\usage{
mm1(drugs, drugdb, admdb, time.points, window, method = "default.doses",
  stash = 0, cap.values = TRUE, verbose = FALSE)
}
\arguments{
\item{drugs}{}

\item{drugdb}{}

\item{admdb}{}

\item{time.points}{}

\item{window}{}

\item{method}{}

\item{stash}{}

\item{cap.values}{}
}
\value{
data.table
}
\description{
Part of the \code{medicinMacro}
}
\author{
Anders Munch
}
