% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mm2.R
\name{mm2}
\alias{mm2}
\title{Calculates exposure periods in a given time span}
\usage{
mm2(
  drugs,
  drugdb,
  admdb,
  periods,
  method = "default.doses",
  stash = 0,
  cap.values = TRUE,
  pre.window = 365.25 * 2,
  print.time = FALSE,
  verbose = FALSE
)
}
\arguments{
\item{drugs}{Relevant drugs}

\item{drugdb}{Drug dataset}

\item{admdb}{Admission dataset}

\item{periods}{Time periods}

\item{method}{Default dose if no other selected}

\item{stash}{Depot}

\item{cap.values}{Overwrite last period with previous}

\item{pre.window}{Time window for prescriptions}

\item{print.time}{Provide time printing}
}
\value{
Exposure period in a given time span
}
\description{
Part of the \code{medicinMacro}
}
\author{
Anders Munch
Helper functions
restricts x to lie within min and max
Find smallest value in the grid, that is larger than x
#' 1: Take out the relevant period and calculate non-hopsitalized days
Probably should be a better way than this?:
Restrict the data
Make periods unique
Only keep latest purchase before start
Add end of each period (= beginning of the next so far)
Add ptt.'s with no purhcases (which might still be hospitalized in the period!):
Also use this place to add 0-exposure periods to patient with a gap from start to first purchase
The following is for handling edge case with hospitalization after start date within first purchase period.
If supplied, calculate non-hospitalized days and add the info
TODO: maybe remove the hospitalization dates outside the periods before forming the cartesian product
Add on the drug info
Overwrite/special case for no purchasers:
#' 2: Calculate end points and daily dose in different cases
NB: non-purchases or total.drug==0 is might be given default doses here, but this is restored below.
The last period should be based on the previous estimate, as we don't know, when the period ends (so can't give a proper daily dose estimate in this case)
So overwrite previous estimate from last period.
Handle case with only one period (set the estimate to the default dosis)
NB: non-purchases or total.drug==0 might be given default doses here, but this is restored below.
Clean up for the edge case of the pre.period: Re-estimate the total drug at start point by substracting the estimated amount of drug spend before start date.
Then overwrite and drop .after.start info for the pre.periods. 
Also overwrite drug info for the cases with no drug left at the start date to not cause confusion. 
From this we calculate other summary measures:
Set n.purchases==0 to estimated dose and days.supply  = 0 (as they are capped or set to default above above):
Also leave out period with 0 days (appears because of the ending dates):
}
