% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/mm2.R
\name{mm2}
\alias{mm2}
\title{Calculates exposure periods in a given time span}
\usage{
mm2(drugs, drugdb, admdb, periods, method = "default.doses", stash = 0,
  cap.values = TRUE, pre.window = 365.25 * 2, print.time = FALSE,
  verbose = FALSE)
}
\arguments{
\item{drugs}{}

\item{drugdb}{}

\item{admdb}{}

\item{periods}{}

\item{method}{}

\item{stash}{}

\item{cap.values}{}

\item{pre.window}{}

\item{print.time}{}
}
\value{
data.table
}
\description{
Part of the \code{medicinMacro}
}
\author{
Anders Munch
}
