% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/drug.R
\name{pack}
\alias{pack}
\title{Add atc codes to object.}
\usage{
d <- dpp()
drug(d, "treatment1") <- atc("A12B")
drug(d, "treatment1") <- pack(c(750, 75), 
                              min = c(250, 25),
                              max = c(1000, 100),
                              def = c(750, 100))
}
\arguments{
\item{value}{The dose sizes corresponding to those found in data. Note that the same value should never be repeated.}

\item{min}{The corresponding minimal doses (same order as 'value').}

\item{max}{The corresponding maximal doses (same order as 'value').}

\item{def}{The corresponding default doses (same order as 'value').}
}
\author{
Helene Charlotte Rytgaard
}

