% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/period.R
\name{period<-}
\alias{period<-}
\title{Set period of time}
\usage{
period(dpp, drug = NULL) <- value
}
\arguments{
\item{dpp}{The preprocessing object}

\item{drug}{If specified, the name (in quotes) of the drug to update the value for. If unspecified, the value is updated for all drugs in the preprocessing object.}

\item{value}{Date vector of start and end date (must be date format).}
}
\description{
Set the relevant period of time.
}
\examples{

d <- dpp()
drug(d, "treatment1") <- atc("A12B")

period(d) <- as.Date(c("1997-01-01", "2012-12-31"))

}
\author{
Helene Charlotte Rytgaard
}

