% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotdata.R
\name{plot.dpp}
\alias{plot.dpp}
\title{Plot function for data preprocessing object}
\usage{
drugdata <- simPrescriptionData(10)
admdata <- simAdmissionData(10)

d <- dpp()

drugdb(d) <- drugdata
admdb(d) <- admdata

plot(d)
}
\arguments{
\item{dpp}{Data preprocessing object}

\item{drug}{If specified, only data for this object is plotted.}
}
\value{
A plot is returned, nothing is changed about the preprocessing object.
}
\description{
Function to plot dates etc
}
\details{
Function to plot dates etc
}
\author{
Helene Charlotte Rytgaard
}

