% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotdata.R
\name{plot.dpp}
\alias{plot.dpp}
\title{Update database}
\usage{
\method{plot}{dpp}(dpp, drug = NULL)
}
\arguments{
\item{dpp}{a}

\item{id}{a}

\item{trace}{a}
}
\description{
Function to plot dates etc
}
\details{
Function to plot dates etc
}
\author{
Helene Charlotte Rytgaard
}

