% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dpp.R
\name{process}
\alias{process}
\title{Data pre-processing}
\usage{
process(x, n = Inf, verbose = TRUE, ...)
}
\arguments{
\item{x}{dpp object}

\item{n}{number of rows from each subject}

\item{verbose}{bla bla?}

\item{...}{not (yet used)}
}
\description{
Pre-processing data
}
\details{
Pre-processing data according to instructions in dpp object
}
