% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/process.R
\name{process}
\alias{process}
\title{Process to estimate prescription lengths and averages}
\usage{
process(dpp, treatments = NULL, id = NULL, trace = FALSE,
  collapse = TRUE)
}
\arguments{
\item{dpp}{Data preprocessing object}

\item{treatments}{If specified, calcations will only be performed
for these treatments.}

\item{id}{If specified, calculations will only be performed for
these individuals.}

\item{trace}{If TRUE, messages are outputted for the user.}

\item{collapse}{Logical. If \code{TRUE}, collapse contiguous
exposure periods with same computed exposure strength to one
and add periods with zero exposure.}
}
\description{
Processing function, to perform calculations according to the data and variables specified in preprocessing object.
}
\examples{
set.seed(05021992)
N=2
da=heaven:::simPrescriptionData(N)
a=heaven:::simAdmissionData(N)
library(heaven)
d=dpp()
drugdb(d) <- da
admdb(d) <- a
drug(d, "drug1") <- atc("A07")
drug(d, "drug1") <- pack(c(200, 400, 500),
                         min = c(100, 200, 250),
                         max = c(300, 800, 1000),
                         def = c(200, 400, 500))
d
maxdepot(d) <- 4000
pwindow(d) <- 2
period(d) <- as.Date(c("1997-01-01", "2012-12-31"))
process(d)
}
\author{
Helene Charlotte Rytgaard
}
