% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/N.R
\name{pwindow<-}
\alias{pwindow<-}
\title{Prescription window}
\usage{
pwindow(dpp, ...) <- value
}
\arguments{
\item{dpp}{a}

\item{...}{a}

\item{value}{a}
}
\description{
Set the number of prescriptions to use in calculations
}
\author{
Helene Charlotte Rytgaard
}

