% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sasWhereConstructors.R
\name{sasDateRange}
\alias{sasDateRange}
\title{sasDateRange}
\usage{
sasDateRange(var.date, date.start, date.end, inclusive = TRUE)
}
\arguments{
\item{var.date}{Name of the date variable in the SAS dataset}

\item{date.start}{The start date of the desired date range in r date format}

\item{date.end}{The end date of the desired date range in r date format}
}
\value{
Function returns a string that can be passed to the "where" statement
  of importSAS. Alternatinvely it can be passed to makeSasWhereStatment
  function to be combined with further SAS where statements
}
\description{
Helper function for making sas "Where" statements for use in importSAS(). Creates date range (inclusive) in SAS formate
}
\examples{
sasDateRange("inddto", as.Date("2015-01-01"), as.Date("2016-01-01"))
}
\author{
Matthew Phelps <mphelps@hjerteforeningen.dk>
}
