% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sasWhereConstructors.R
\name{sasWhereConstructor}
\alias{sasWhereConstructor}
\title{sasWhereConstructor}
\usage{
sasWhereConstructor(..., collapse = " AND ")
}
\arguments{
\item{...}{A vector of valid SAS where statements.}

\item{collapse}{Argment passed to paste0. default is \code{" AND "}.}
}
\value{
Returns a string that can be passed to the "where" argument of importSAS
}
\description{
Simple helper function that combines multiple SAS where statements with "AND"
}
\author{
Matthew Phelps <mphelps@hjerteforeningen.dk>
}
