% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dpp.R
\name{selector}
\alias{selector}
\title{Select variables}
\usage{
selector(data, var, by = "pnr", search.term, sortkey, period = NULL,
  backward = NULL, forward = NULL, select = "first", name = NULL,
  collect = NULL, ...)
}
\arguments{
\item{data}{Optional. The name of a registered data set in which to evaluate the instructions.}

\item{var}{Variable to filter with. Could be \code{"diag"} when data is set to \code{"lpr"}.}

\item{by}{Character vector. Specifying subsets in which \code{value} is evaluated. E.g., if set to \code{"pnr"} then \code{value} is applied to the data of each subject separately.}

\item{search.term}{Expression to filter with. Could be (partial) ATC or (partial) ICD codes pasted together with \code{|}.}

\item{sortkey}{Date variable for sorting in time. Could be \code{"uddto"} when data is \code{"lpr"}}

\item{period}{Calendar time period}

\item{backward}{Object from which to look backwards in time.}

\item{forward}{Object from which to look forwards in time.}

\item{select}{Either a character ("first", "last", "any", "count") or a function.}

\item{name}{Name or vector of names for the new variable(s).}

\item{collect}{Named vector or list with variables to collect}

\item{...}{Arguments passed to the result}
}
\description{
All purpose function to contstruct instructions for inclusion, exclusion and variables
}
\details{
Returns instructions (a function) which creates variables from one or all of the regis data sets.
}
