% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/simAdmissionData.R
\name{simAdmissionData}
\alias{simAdmissionData}
\title{Admission data simulation function}
\usage{
simAdmissionData(n, m = 5, startDate = "1995-01-01", longformat = TRUE)
}
\arguments{
\item{n}{Number of patients}

\item{m}{Maximal number of admission dates per patient}
}
\description{
Simulate admission data
}
\details{
Simulate admission data alike the Danish medical registry
}
\examples{

set.seed(1)
simAdmissionData(1)
set.seed(2)
simAdmissionData(2)

}
\author{
Helene Charlotte Rytgaard \email{hely@sund.ku.dk}
}

