% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/standardize.rate.R
\name{standardize.rate}
\alias{standardize.rate}
\title{Standardize proportions and absolute risks to a given age
    distribution}
\usage{
standardize.rate(x, age = "agegroups", exposure, by,
  standardize.to = "ref.level", data, method = "f", level = 0.95,
  crude = TRUE, ...)
}
\arguments{
\item{age}{Name of categorical age variable}

\item{by}{Name(s) of categorical strata variable(s)}

\item{standardize.to}{what population to use for standardization.}

\item{data}{Data set which contains all variables}

\item{level}{Confidence level}

\item{...}{Not (yet) used}

\item{vars}{Names of variable(s) which contain the rate(s) to be
standardized.}
}
\value{
Data table with standardized rates
}
\description{
Standardize proportions and absolute risks to a given age distribution
}
\details{
Standardize proportions and absolute risks to a given age
distribution
}
\examples{
library(riskRegression)
set.seed(84)
n=160
d <- data.table(e1=rpois(n,lambda=9),
                rt1=rpois(n,lambda=1880),
                e2=rpois(n,lambda=123),
                rt2=rpois(n,lambda=80000))
d[,agegroups:=factor(rep(c("40-50","45-50","50-55","55-60","60-65","65-70","70-75","75-80"),n/8))]
d[,sex:=factor(rep(c("f","m"),c(n/2,n/2)))]
d[,year:=rep(2001:2010,n/10)]
D=d[,.(e1=sum(e1),rt1=sum(rt1),e2=sum(e2),rt2=sum(rt2)),by=c("sex","agegroups")]
D[sex=="m",e1:=e1+rpois(.N,lambda=as.numeric(agegroups)*17)]
D[sex=="m",rt1:=rt1-rpois(.N,lambda=as.numeric(agegroups)*1600)]
standardize.rate(x=list(c("e1","rt1")),
                 age="agegroups",exposure="sex",data=D,standardize.to="f")
standardize.rate(x=list(c("e1","rt1")),
                 age="agegroups",exposure="sex",data=D,standardize.to="m")
standardize.rate(x=list(c("e1","rt1")),
                 age="agegroups",exposure="sex",data=D,standardize.to="mean")
standardize.rate(x=list(c("e1","rt1"),c("e2","rt2")),
                 age="agegroups",exposure="sex",data=d)

}
\seealso{
standardize.prodlim standardize.proportion
}
\author{
Thomas A. Gerds <tag@biostat.ku.dk>
}
