% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xrecepter.R
\name{xrecepter}
\alias{xrecepter}
\title{wrapper for medicin macro}
\usage{
xrecepter(drug, adm, atc, npre, dose, min, max, def, maxdepot, period,
  name, user, server = "doob", sas = "/usr/local/bin/sas ",
  macro = "~/research/SoftWare/heaven/sas/medicin_macro_version_31.sas",
  proFile = "/tmp/med-macro-call.sas",
  drugFile = "/tmp/med-macro-drug.csv",
  admFile = "/tmp/med-macro-adm.csv",
  outFile = "/tmp/med-macro-out.csv", logfile = "/tmp/med-macro.log",
  lstfile = "/tmp/med-macro.lst")
}
\arguments{
\item{drug}{drug data}

\item{adm}{admission data}

\item{atc}{atc kode specifying drug of interest}

\item{name}{name of drug}

\item{sas}{path to sas program}

\item{macro}{path to sas macro}

\item{proFile}{path to pro}

\item{drugFile}{path to drug}

\item{admFile}{path to adm}

\item{outFile}{path to out}

\item{logfile}{path to log}
}
\description{
wrapper for medicin macro
}
\details{
wrapper for medicin macro
}
\examples{
library(heaven)
ppi=read.table("~/research/SoftWare/heaven/data/PPI.csv",sep=";",header=TRUE)
adm=read.table("~/research/SoftWare/heaven/data/admData.csv",sep=";",header=TRUE)
u=medicinMacro(drug=ppi,adm=adm,atc="A02BC02",name="omeprazol",npre=5,dose=c(10, 20, 40, 40),min=c(10, 20, 40, 40),max=c(20, 40, 60, 80),def=c(10, 20, 40, 40),maxdepot=4000,period=c("'01jan1997'd","'31dec2012'd"),sas="/usr/local/bin/sas",server="doob",user="grb615")
v=medicinMacro(drug=ppi,adm=adm,atc="A02BC02",name="omeprazol",npre=5,dose=c(10, 20, 40, 40),min=c(10, 20, 40, 40),max=c(20, 40, 60, 80),def=c(20, 40, 60, 10),maxdepot=4000,period=c("'01jan1997'd","'31dec2012'd"),sas="/usr/local/bin/sas",server="doob",user="grb615")

set.seed(05021992)
N=100
da=heaven:::simPrescriptionData(N)
a=heaven:::simAdmissionData(N)
system.time(w <- medicinMacro(drug=da,adm=a,atc="A07",name="drug1",npre=5,dose=c(200, 400, 500, 500),min=c(100, 200, 250, 250),max=c(300, 800, 1000, 1000),def=c(200, 400,500,500),maxdepot=4000,period=c("'01jan1997'd","'31dec2012'd"),sas="/usr/local/bin/sas",server="doob",user="grb615"))

library(heaven)
d=dpp()
drugdb(d) <- (da)
admdb(d) <- (a)

drug(d, "drug1") <- atc("A07")
drug(d, "drug1") <- pack(c(200, 400, 500),
                         min = c(100, 200, 250),
                         max = c(300, 800, 1000),
                         def = c(200, 400, 500))
d
maxdepot(d) <- 4000
pwindow(d) <- 2
period(d) <- as.Date(c("1997-01-01", "2012-12-31"))
system.time(pd <- process(d,out_data=TRUE))
x=pd$drug1[pd$drug1$X!=0,]
for(i in unique(da$pnr)){try(print(cbind(w[w$pnr==i,],x[x$id==i,])))}

}
\author{
Thomas A. Gerds <tag@biostat.ku.dk>
}
