% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/xrecepter.R
\name{xrecepter}
\alias{xrecepter}
\title{wrapper for medicin macro}
\usage{
xrecepter(drugdb, adm, atc, window = 5, value, min, max, def, maxdepot,
  period = c("'01jan1995'd", "'31dec2015'd"), name = "adrug", user, wd,
  sas.program = "/usr/local/bin/sas ", sas.switches, sas.runner,
  macro = "~/research/SoftWare/heaven/sas/medicin_macro_version_31.sas",
  save.tmp = FALSE, verbose = TRUE, server = "doob",
  remote = FALSE, remote.home)
}
\arguments{
\item{drugdb}{drugdb data}

\item{adm}{admission data}

\item{atc}{atc kode specifying drugdb of interest}

\item{window}{number of purchases that enter the calculations}

\item{value}{is a vector of strengths of one
unit (e.g., pill) of the drug. The vector should have one such strength for
each of the different packages that occur in the data.}

\item{min}{is a vector of the same length as \code{value} where each element is the
assumed minimum total dosis that a patient can consume on one day. For example, if the
value is 50 mg and the pills of this strength can be cut (in halves) the minimum
total dosis is 25 mg.}

\item{max}{is a vector of the same length as \code{value} where each element is the
assumed maximum total dosis that a patient can consume on one day. For example, if the
value is 50 mg and one can at most consume 4 pills a day the maximum
total dosis is 200 mg.}

\item{def}{is a vector of the same length as \code{value} where each element is the
assumed default dosis that an average patient would consume on one day. For example,
if the value is 50 mg and usually a patient would consume 2 pills the default is 100 mg.}

\item{maxdepot}{Maximum number of units that a patient can stash}

\item{period}{Vector of two dates that SAS can eat: c("'01jan1995'd","'31dec2015'd").
Can also be numeric values: number of days since 1960-01-01.}

\item{name}{working name of output file (only relevant when save.tmp is set)}

\item{user}{user name on remote (unix like) machine}

\item{wd}{working directory}

\item{sas.program}{path to sas program}

\item{sas.switches}{On linux this defaults to {""} on any other
system to \code{"-batch -nosplash -noenhancededitor -sysin"}}

\item{sas.runner}{How sas is invoked. On linux this defaults to
\code{"system"} on any other system to \code{"shell"}.}

\item{macro}{path to sas macro}

\item{save.tmp}{if TRUE do not delete the temporary SAS files}

\item{verbose}{like bla bla?}

\item{server}{name of remote (unix like) machine}

\item{remote}{if TRUE run on a remote (unix like) machine}

\item{remote.home}{home folder on remote (unix like) machine}
}
\description{
wrapper for medicin macro
}
\details{
wrapper for medicin macro
}
\examples{
library(heaven)
\dontrun{
 library(heaven)
library(data.table)
set.seed(05021992)
N=18
packs = list("R03AK11"=list(c(10,1)))
lmdb=simPrescriptionData(N,packages=packs,max.packages=1)
lpr=simAdmissionData(N)
## very simple data 
lmdb <- lmdb[c(1,4,18)]
R03 = list(atc=c("R03AK11"),
          maxdepot=100,
          period=as.Date(c("1995-01-01", "2012-12-31")),
           prescriptionwindow=2,
          doses=list(value=c(5,10),
                     min = c(.5,.5),
                     max = c(2,2),
                     def = c(1,1)))
xrecepter(drugdb=lmdb,adm=lpr,window=5,remote=FALSE,wd="~/tmp",save.tmp=0L,
          value=rep(R03$doses$value,2),min=rep(R03$doses$min,2),max=rep(R03$doses$max,2),
          def=rep(R03$doses$def,2),maxdepot=100,period=c(12784,20089),atc="R03AK11",
          name="blaupill",sas.program="/usr/local/bin/sas",
          verbose=FALSE)
} 

}
\author{
Thomas A. Gerds <tag@biostat.ku.dk>
}
